Backport WT-4723 (Restructure the reconciliation code) to 4.0.
