+++ src/btree/bt_sync.c
@@ -143,7 +143,7 @@ __sync_file(WT_SESSION_IMPL *session, WT_CACHE_OP syncop)
 	 * checkpoint, the on-disk version is correct.  If the truncate is
 	 * visible, we skip over the child page when writing its parent.  We
 	 * check whether a truncate is visible in the checkpoint as part of
-	 * reconciling internal pages (specifically in __rec_child_modify).
+	 * reconciling internal pages (specifically in __wt_rec_child_modify).
 	 */
 	LF_SET(WT_READ_DELETED_SKIP);
