+++ src/reconcile/rec_write.c
@@ -8,319 +8,24 @@
 
 #include "wt_internal.h"
 
-struct __rec_chunk;		typedef struct __rec_chunk WT_CHUNK;
-struct __rec_dictionary;	typedef struct __rec_dictionary WT_DICTIONARY;
-struct __rec_kv;		typedef struct __rec_kv WT_KV;
-
-/*
- * Reconciliation is the process of taking an in-memory page, walking each entry
- * in the page, building a backing disk image in a temporary buffer representing
- * that information, and writing that buffer to disk.  What could be simpler?
- *
- * WT_RECONCILE --
- *	Information tracking a single page reconciliation.
- */
-typedef struct {
-	WT_REF  *ref;			/* Page being reconciled */
-	WT_PAGE *page;
-	uint32_t flags;			/* Caller's configuration */
-
-	/*
-	 * Track start/stop write generation to decide if all changes to the
-	 * page are written.
-	 */
-	uint32_t orig_write_gen;
-
-	/*
-	 * Track start/stop checkpoint generations to decide if lookaside table
-	 * records are correct.
-	 */
-	uint64_t orig_btree_checkpoint_gen;
-	uint64_t orig_txn_checkpoint_gen;
-
-	/*
-	 * Track the oldest running transaction and whether to skew lookaside
-	 * to the newest update.
-	 */
-	bool las_skew_newest;
-	uint64_t last_running;
-
-	/* Track the page's min/maximum transactions. */
-	uint64_t max_txn;
-	wt_timestamp_t max_timestamp;
-
-	/* Lookaside boundary tracking. */
-	uint64_t unstable_txn;
-	wt_timestamp_t unstable_timestamp;
-
-	u_int updates_seen;		/* Count of updates seen. */
-	u_int updates_unstable;		/* Count of updates not visible_all. */
-
-	bool update_uncommitted;	/* An update was uncommitted */
-	bool update_used;		/* An update could be used */
-
-	/*
-	 * When we can't mark the page clean (for example, checkpoint found some
-	 * uncommitted updates), there's a leave-dirty flag.
-	 */
-	bool leave_dirty;
-
-	/*
-	 * Track if reconciliation has seen any overflow items.  If a leaf page
-	 * with no overflow items is written, the parent page's address cell is
-	 * set to the leaf-no-overflow type.  This means we can delete the leaf
-	 * page without reading it because we don't have to discard any overflow
-	 * items it might reference.
-	 *
-	 * The test test is per-page reconciliation, that is, once we see an
-	 * overflow item on the page, all subsequent leaf pages written for the
-	 * page will not be leaf-no-overflow type, regardless of whether or not
-	 * they contain overflow items.  In other words, leaf-no-overflow is not
-	 * guaranteed to be set on every page that doesn't contain an overflow
-	 * item, only that if it is set, the page contains no overflow items.
-	 * XXX
-	 * This was originally done because raw compression couldn't do better,
-	 * now that raw compression has been removed, we should do better.
-	 */
-	bool	ovfl_items;
-
-	/*
-	 * Track if reconciliation of a row-store leaf page has seen empty (zero
-	 * length) values.  We don't write out anything for empty values, so if
-	 * there are empty values on a page, we have to make two passes over the
-	 * page when it's read to figure out how many keys it has, expensive in
-	 * the common case of no empty values and (entries / 2) keys.  Likewise,
-	 * a page with only empty values is another common data set, and keys on
-	 * that page will be equal to the number of entries.  In both cases, set
-	 * a flag in the page's on-disk header.
-	 *
-	 * The test is per-page reconciliation as described above for the
-	 * overflow-item test.
-	 */
-	bool	all_empty_value, any_empty_value;
-
-	/*
-	 * Reconciliation gets tricky if we have to split a page, which happens
-	 * when the disk image we create exceeds the page type's maximum disk
-	 * image size.
-	 *
-	 * First, the target size of the page we're building.
-	 */
-	uint32_t page_size;		/* Page size */
-
-	/*
-	 * Second, the split size: if we're doing the page layout, split to a
-	 * smaller-than-maximum page size when a split is required so we don't
-	 * repeatedly split a packed page.
-	 */
-	uint32_t split_size;		/* Split page size */
-	uint32_t min_split_size;	/* Minimum split page size */
-
-	/*
-	 * We maintain two split chunks in the memory during reconciliation to
-	 * be written out as pages. As we get to the end of the data, if the
-	 * last one turns out to be smaller than the minimum split size, we go
-	 * back into the penultimate chunk and split at this minimum split size
-	 * boundary. This moves some data from the penultimate chunk to the last
-	 * chunk, hence increasing the size of the last page written without
-	 * decreasing the penultimate page size beyond the minimum split size.
-	 * For this reason, we maintain an expected split percentage boundary
-	 * and a minimum split percentage boundary.
-	 *
-	 * Chunks are referenced by current and previous pointers. In case of a
-	 * split, previous references the first chunk and current switches to
-	 * the second chunk. If reconciliation generates more split chunks, the
-	 * the previous chunk is written to the disk and current and previous
-	 * swap.
-	 */
-	struct __rec_chunk {
-		/*
-		 * The recno and entries fields are the starting record number
-		 * of the split chunk (for column-store splits), and the number
-		 * of entries in the split chunk.
-		 *
-		 * The key for a row-store page; no column-store key is needed
-		 * because the page's recno, stored in the recno field, is the
-		 * column-store key.
-		 */
-		uint32_t entries;
-		uint64_t recno;
-		WT_ITEM  key;
-
-		uint32_t min_entries;
-		uint64_t min_recno;
-		WT_ITEM  min_key;
-
-		/* Minimum split-size boundary buffer offset. */
-		size_t   min_offset;
-
-		WT_ITEM image;				/* disk-image */
-	} chunkA, chunkB, *cur_ptr, *prev_ptr;
-
-	/*
-	 * We track current information about the current record number, the
-	 * number of entries copied into the disk image buffer, where we are
-	 * in the buffer, and how much memory remains. Those values are
-	 * packaged here rather than passing pointers to stack locations
-	 * around the code.
-	 */
-	uint64_t recno;			/* Current record number */
-	uint32_t entries;		/* Current number of entries */
-	uint8_t *first_free;		/* Current first free byte */
-	size_t	 space_avail;		/* Remaining space in this chunk */
-	/* Remaining space in this chunk to put a minimum size boundary */
-	size_t	 min_space_avail;
-
-	/*
-	 * Saved update list, supporting the WT_REC_UPDATE_RESTORE and
-	 * WT_REC_LOOKASIDE configurations. While reviewing updates for each
-	 * page, we save WT_UPDATE lists here, and then move them to per-block
-	 * areas as the blocks are defined.
-	 */
-	WT_SAVE_UPD *supd;		/* Saved updates */
-	uint32_t     supd_next;
-	size_t	     supd_allocated;
-	size_t       supd_memsize;	/* Size of saved update structures */
-
-	/* List of pages we've written so far. */
-	WT_MULTI *multi;
-	uint32_t  multi_next;
-	size_t	  multi_allocated;
-
-	/*
-	 * Root pages are written when wrapping up the reconciliation, remember
-	 * the image we're going to write.
-	 */
-	WT_ITEM *wrapup_checkpoint;
-	bool	 wrapup_checkpoint_compressed;
-
-	/*
-	 * We don't need to keep the 0th key around on internal pages, the
-	 * search code ignores them as nothing can sort less by definition.
-	 * There's some trickiness here, see the code for comments on how
-	 * these fields work.
-	 */
-	bool	cell_zero;		/* Row-store internal page 0th key */
-
-	/*
-	 * We calculate checksums to find previously written identical blocks,
-	 * but once a match fails during an eviction, there's no point trying
-	 * again.
-	 */
-	bool	evict_matching_checksum_failed;
-
-	/*
-	 * WT_DICTIONARY --
-	 *	We optionally build a dictionary of values for leaf pages. Where
-	 * two value cells are identical, only write the value once, the second
-	 * and subsequent copies point to the original cell. The dictionary is
-	 * fixed size, but organized in a skip-list to make searches faster.
-	 */
-	struct __rec_dictionary {
-		uint64_t hash;				/* Hash value */
-		uint32_t offset;			/* Matching cell */
-
-		u_int depth;				/* Skiplist */
-		WT_DICTIONARY *next[0];
-	} **dictionary;					/* Dictionary */
-	u_int dictionary_next, dictionary_slots;	/* Next, max entries */
-							/* Skiplist head. */
-	WT_DICTIONARY *dictionary_head[WT_SKIP_MAXDEPTH];
-
-	/*
-	 * WT_KV--
-	 *	An on-page key/value item we're building.
-	 */
-	struct __rec_kv {
-		WT_ITEM	 buf;		/* Data */
-		WT_CELL	 cell;		/* Cell and cell's length */
-		size_t cell_len;
-		size_t len;		/* Total length of cell + data */
-	} k, v;				/* Key/Value being built */
-
-	WT_ITEM *cur, _cur;		/* Key/Value being built */
-	WT_ITEM *last, _last;		/* Last key/value built */
-
-	bool key_pfx_compress;		/* If can prefix-compress next key */
-	bool key_pfx_compress_conf;	/* If prefix compression configured */
-	bool key_sfx_compress;		/* If can suffix-compress next key */
-	bool key_sfx_compress_conf;	/* If suffix compression configured */
-
-	bool is_bulk_load;		/* If it's a bulk load */
-
-	WT_SALVAGE_COOKIE *salvage;	/* If it's a salvage operation */
-
-	bool cache_write_lookaside;	/* Used the lookaside table */
-	bool cache_write_restore;	/* Used update/restoration */
-
-	uint32_t tested_ref_state;	/* Debugging information */
-
-	/*
-	 * XXX
-	 * In the case of a modified update, we may need a copy of the current
-	 * value as a set of bytes. We call back into the btree code using a
-	 * fake cursor to do that work. This a layering violation and fragile,
-	 * we need a better solution.
-	 */
-	WT_CURSOR_BTREE update_modify_cbt;
-} WT_RECONCILE;
-
-#define	WT_CROSSING_MIN_BND(r, next_len)				\
-	((r)->cur_ptr->min_offset == 0 &&				\
-	    (next_len) > (r)->min_space_avail)
-#define	WT_CROSSING_SPLIT_BND(r, next_len) ((next_len) > (r)->space_avail)
-#define	WT_CHECK_CROSSING_BND(r, next_len)				\
-	(WT_CROSSING_MIN_BND(r, next_len) || WT_CROSSING_SPLIT_BND(r, next_len))
-
-static void __rec_cell_build_addr(WT_SESSION_IMPL *,
-		WT_RECONCILE *, const void *, size_t, u_int, uint64_t);
-static int  __rec_cell_build_int_key(WT_SESSION_IMPL *,
-		WT_RECONCILE *, const void *, size_t, bool *);
-static int  __rec_cell_build_leaf_key(WT_SESSION_IMPL *,
-		WT_RECONCILE *, const void *, size_t, bool *);
-static int  __rec_cell_build_ovfl(WT_SESSION_IMPL *,
-		WT_RECONCILE *, WT_KV *, uint8_t, uint64_t);
-static int  __rec_cell_build_val(WT_SESSION_IMPL *,
-		WT_RECONCILE *, const void *, size_t, uint64_t);
 static void __rec_cleanup(WT_SESSION_IMPL *, WT_RECONCILE *);
-static int  __rec_col_fix(WT_SESSION_IMPL *, WT_RECONCILE *, WT_REF *);
-static int  __rec_col_fix_slvg(WT_SESSION_IMPL *,
-		WT_RECONCILE *, WT_REF *, WT_SALVAGE_COOKIE *);
-static int  __rec_col_int(WT_SESSION_IMPL *, WT_RECONCILE *, WT_REF *);
-static int  __rec_col_merge(WT_SESSION_IMPL *, WT_RECONCILE *, WT_PAGE *);
-static int  __rec_col_var(WT_SESSION_IMPL *,
-		WT_RECONCILE *, WT_REF *, WT_SALVAGE_COOKIE *);
-static int  __rec_col_var_helper(WT_SESSION_IMPL *, WT_RECONCILE *,
-		WT_SALVAGE_COOKIE *, WT_ITEM *, bool, uint8_t, uint64_t);
 static int  __rec_destroy_session(WT_SESSION_IMPL *);
 static int  __rec_init(WT_SESSION_IMPL *,
 		WT_REF *, uint32_t, WT_SALVAGE_COOKIE *, void *);
 static int  __rec_las_wrapup(WT_SESSION_IMPL *, WT_RECONCILE *);
 static int  __rec_las_wrapup_err(WT_SESSION_IMPL *, WT_RECONCILE *);
 static int  __rec_root_write(WT_SESSION_IMPL *, WT_PAGE *, uint32_t);
-static int  __rec_row_int(WT_SESSION_IMPL *, WT_RECONCILE *, WT_PAGE *);
-static int  __rec_row_leaf(WT_SESSION_IMPL *,
-		WT_RECONCILE *, WT_PAGE *, WT_SALVAGE_COOKIE *);
-static int  __rec_row_leaf_insert(
-		WT_SESSION_IMPL *, WT_RECONCILE *, WT_INSERT *);
-static int  __rec_row_merge(WT_SESSION_IMPL *, WT_RECONCILE *, WT_PAGE *);
 static int  __rec_split_discard(WT_SESSION_IMPL *, WT_PAGE *);
 static int  __rec_split_row_promote(
 		WT_SESSION_IMPL *, WT_RECONCILE *, WT_ITEM *, uint8_t);
-static int  __rec_split_write(
-		WT_SESSION_IMPL *, WT_RECONCILE *, WT_CHUNK *, WT_ITEM *, bool);
+static int  __rec_split_write(WT_SESSION_IMPL *,
+		WT_RECONCILE *, WT_REC_CHUNK *, WT_ITEM *, bool);
 static int  __rec_write_check_complete(
 		WT_SESSION_IMPL *, WT_RECONCILE *, int, bool *);
 static void __rec_write_page_status(WT_SESSION_IMPL *, WT_RECONCILE *);
 static int  __rec_write_wrapup(WT_SESSION_IMPL *, WT_RECONCILE *, WT_PAGE *);
 static int  __rec_write_wrapup_err(
 		WT_SESSION_IMPL *, WT_RECONCILE *, WT_PAGE *);
-
-static void __rec_dictionary_free(WT_SESSION_IMPL *, WT_RECONCILE *);
-static int  __rec_dictionary_init(WT_SESSION_IMPL *, WT_RECONCILE *, u_int);
-static int  __rec_dictionary_lookup(
-		WT_SESSION_IMPL *, WT_RECONCILE *, WT_KV *, WT_DICTIONARY **);
-static void __rec_dictionary_reset(WT_RECONCILE *);
 
 /*
  * __wt_reconcile --
@@ -435,23 +140,23 @@ __wt_reconcile(WT_SESSION_IMPL *session, WT_REF *ref,
 	switch (page->type) {
 	case WT_PAGE_COL_FIX:
 		if (salvage != NULL)
-			ret = __rec_col_fix_slvg(session, r, ref, salvage);
+			ret = __wt_rec_col_fix_slvg(session, r, ref, salvage);
 		else
-			ret = __rec_col_fix(session, r, ref);
+			ret = __wt_rec_col_fix(session, r, ref);
 		break;
 	case WT_PAGE_COL_INT:
 		WT_WITH_PAGE_INDEX(session,
-		    ret = __rec_col_int(session, r, ref));
+		    ret = __wt_rec_col_int(session, r, ref));
 		break;
 	case WT_PAGE_COL_VAR:
-		ret = __rec_col_var(session, r, ref, salvage);
+		ret = __wt_rec_col_var(session, r, ref, salvage);
 		break;
 	case WT_PAGE_ROW_INT:
 		WT_WITH_PAGE_INDEX(session,
-		    ret = __rec_row_int(session, r, page));
+		    ret = __wt_rec_row_int(session, r, page));
 		break;
 	case WT_PAGE_ROW_LEAF:
-		ret = __rec_row_leaf(session, r, page, salvage);
+		ret = __wt_rec_row_leaf(session, r, page, salvage);
 		break;
 	default:
 		ret = __wt_illegal_value(session, page->type);
@@ -965,9 +670,9 @@ __rec_init(WT_SESSION_IMPL *session,
 	 * Sanity check the size: 100 slots is the smallest dictionary we use.
 	 */
 	if (btree->dictionary != 0 && btree->dictionary > r->dictionary_slots)
-		WT_RET(__rec_dictionary_init(session,
+		WT_RET(__wt_rec_dictionary_init(session,
 		    r, btree->dictionary < 100 ? 100 : btree->dictionary));
-	__rec_dictionary_reset(r);
+	__wt_rec_dictionary_reset(r);
 
 	/*
 	 * Prefix compression discards repeated prefix bytes from row-store leaf
@@ -1059,7 +764,7 @@ __rec_destroy(WT_SESSION_IMPL *session, void *reconcilep)
 
 	__wt_free(session, r->supd);
 
-	__rec_dictionary_free(session, r);
+	__wt_rec_dictionary_free(session, r);
 
 	__wt_buf_free(session, &r->k.buf);
 	__wt_buf_free(session, &r->v.buf);
@@ -1081,914 +786,6 @@ __rec_destroy_session(WT_SESSION_IMPL *session)
 	__rec_destroy(session, &session->reconcile);
 	return (0);
 }
-
-/*
- * __rec_update_save --
- *	Save a WT_UPDATE list for later restoration.
- */
-static int
-__rec_update_save(WT_SESSION_IMPL *session, WT_RECONCILE *r,
-    WT_INSERT *ins, void *ripcip, WT_UPDATE *onpage_upd, size_t upd_memsize)
-{
-	WT_RET(__wt_realloc_def(
-	    session, &r->supd_allocated, r->supd_next + 1, &r->supd));
-	r->supd[r->supd_next].ins = ins;
-	r->supd[r->supd_next].ripcip = ripcip;
-	r->supd[r->supd_next].onpage_upd = onpage_upd;
-	++r->supd_next;
-	r->supd_memsize += upd_memsize;
-	return (0);
-}
-
-/*
- * __rec_append_orig_value --
- *	Append the key's original value to its update list.
- */
-static int
-__rec_append_orig_value(WT_SESSION_IMPL *session,
-    WT_PAGE *page, WT_UPDATE *upd, WT_CELL_UNPACK *unpack)
-{
-	WT_DECL_ITEM(tmp);
-	WT_DECL_RET;
-	WT_UPDATE *append;
-	size_t size;
-
-	/* Done if at least one self-contained update is globally visible. */
-	for (;; upd = upd->next) {
-		if (WT_UPDATE_DATA_VALUE(upd) &&
-		    __wt_txn_upd_visible_all(session, upd))
-			return (0);
-
-		/* Add the original value after birthmarks. */
-		if (upd->type == WT_UPDATE_BIRTHMARK) {
-			WT_ASSERT(session, unpack != NULL &&
-			    unpack->type != WT_CELL_DEL);
-			break;
-		}
-
-		/* Leave reference at the last item in the chain. */
-		if (upd->next == NULL)
-			break;
-	}
-
-	/*
-	 * We need the original on-page value for some reader: get a copy and
-	 * append it to the end of the update list with a transaction ID that
-	 * guarantees its visibility.
-	 *
-	 * If we don't have a value cell, it's an insert/append list key/value
-	 * pair which simply doesn't exist for some reader; place a deleted
-	 * record at the end of the update list.
-	 */
-	append = NULL;			/* -Wconditional-uninitialized */
-	size = 0;			/* -Wconditional-uninitialized */
-	if (unpack == NULL || unpack->type == WT_CELL_DEL)
-		WT_RET(__wt_update_alloc(session,
-		    NULL, &append, &size, WT_UPDATE_TOMBSTONE));
-	else {
-		WT_RET(__wt_scr_alloc(session, 0, &tmp));
-		WT_ERR(__wt_page_cell_data_ref(session, page, unpack, tmp));
-		WT_ERR(__wt_update_alloc(
-		    session, tmp, &append, &size, WT_UPDATE_STANDARD));
-	}
-
-	/*
-	 * If we're saving the original value for a birthmark, transfer over
-	 * the transaction ID and clear out the birthmark update.
-	 *
-	 * Else, set the entry's transaction information to the lowest possible
-	 * value. Cleared memory matches the lowest possible transaction ID and
-	 * timestamp, do nothing.
-	 */
-	if (upd->type == WT_UPDATE_BIRTHMARK) {
-		append->txnid = upd->txnid;
-		append->timestamp = upd->timestamp;
-		append->next = upd->next;
-	}
-
-	/* Append the new entry into the update list. */
-	WT_PUBLISH(upd->next, append);
-	__wt_cache_page_inmem_incr(session, page, size);
-
-	if (upd->type == WT_UPDATE_BIRTHMARK) {
-		upd->type = WT_UPDATE_STANDARD;
-		upd->txnid = WT_TXN_ABORTED;
-	}
-
-err:	__wt_scr_free(session, &tmp);
-	return (ret);
-}
-
-/*
- * __rec_txn_read --
- *	Return the update in a list that should be written (or NULL if none can
- *	be written).
- */
-static int
-__rec_txn_read(WT_SESSION_IMPL *session, WT_RECONCILE *r,
-    WT_INSERT *ins, void *ripcip, WT_CELL_UNPACK *vpack,
-    bool *upd_savedp, WT_UPDATE **updp)
-{
-	WT_PAGE *page;
-	WT_UPDATE *first_ts_upd, *first_txn_upd, *first_upd, *upd;
-	wt_timestamp_t timestamp;
-	size_t upd_memsize;
-	uint64_t max_txn, txnid;
-	bool all_visible, prepared, skipped_birthmark, uncommitted, upd_saved;
-
-	if (upd_savedp != NULL)
-		*upd_savedp = false;
-	*updp = NULL;
-
-	page = r->page;
-	first_ts_upd = first_txn_upd = NULL;
-	upd_memsize = 0;
-	max_txn = WT_TXN_NONE;
-	prepared = skipped_birthmark = uncommitted = upd_saved = false;
-
-	/*
-	 * If called with a WT_INSERT item, use its WT_UPDATE list (which must
-	 * exist), otherwise check for an on-page row-store WT_UPDATE list
-	 * (which may not exist). Return immediately if the item has no updates.
-	 */
-	if (ins != NULL)
-		first_upd = ins->upd;
-	else if ((first_upd = WT_ROW_UPDATE(page, ripcip)) == NULL)
-		return (0);
-
-	for (upd = first_upd; upd != NULL; upd = upd->next) {
-		if ((txnid = upd->txnid) == WT_TXN_ABORTED)
-			continue;
-
-		++r->updates_seen;
-		upd_memsize += WT_UPDATE_MEMSIZE(upd);
-
-		/*
-		 * Track the first update in the chain that is not aborted and
-		 * the maximum transaction ID.
-		 */
-		if (first_txn_upd == NULL)
-			first_txn_upd = upd;
-
-		/* Track the largest transaction ID seen. */
-		if (WT_TXNID_LT(max_txn, txnid))
-			max_txn = txnid;
-
-		/*
-		 * Check whether the update was committed before reconciliation
-		 * started.  The global commit point can move forward during
-		 * reconciliation so we use a cached copy to avoid races when a
-		 * concurrent transaction commits or rolls back while we are
-		 * examining its updates. As prepared transaction id's are
-		 * globally visible, need to check the update state as well.
-		 */
-		if (F_ISSET(r, WT_REC_EVICT)) {
-			if (upd->prepare_state == WT_PREPARE_LOCKED ||
-			    upd->prepare_state == WT_PREPARE_INPROGRESS)
-				prepared = true;
-
-			if (F_ISSET(r, WT_REC_VISIBLE_ALL) ?
-			    WT_TXNID_LE(r->last_running, txnid) :
-			    !__txn_visible_id(session, txnid))
-				uncommitted = r->update_uncommitted = true;
-
-		       if (prepared || uncommitted)
-			       continue;
-		}
-
-		/* Track the first update with non-zero timestamp. */
-		if (first_ts_upd == NULL && upd->timestamp != 0)
-			first_ts_upd = upd;
-
-		/*
-		 * Find the first update we can use.
-		 *
-		 * Update/restore eviction can handle any update (including
-		 * uncommitted updates).  Lookaside eviction can save any
-		 * committed update.  Regular eviction checks that the maximum
-		 * transaction ID and timestamp seen are stable.
-		 *
-		 * Lookaside and update/restore eviction try to choose the same
-		 * version as a subsequent checkpoint, so that checkpoint can
-		 * skip over pages with lookaside entries.  If the application
-		 * has supplied a stable timestamp, we assume (a) that it is
-		 * old, and (b) that the next checkpoint will use it, so we wait
-		 * to see a stable update.  If there is no stable timestamp, we
-		 * assume the next checkpoint will write the most recent version
-		 * (but we save enough information that checkpoint can fix
-		 * things up if we choose an update that is too new).
-		 */
-		if (*updp == NULL && r->las_skew_newest)
-			*updp = upd;
-
-		if (F_ISSET(r, WT_REC_VISIBLE_ALL) ?
-		    !__wt_txn_upd_visible_all(session, upd) :
-		    !__wt_txn_upd_visible(session, upd)) {
-			if (F_ISSET(r, WT_REC_EVICT))
-				++r->updates_unstable;
-
-			/*
-			 * Rare case: when applications run at low isolation
-			 * levels, update/restore eviction may see a stable
-			 * update followed by an uncommitted update.  Give up
-			 * in that case: we need to discard updates from the
-			 * stable update and older for correctness and we can't
-			 * discard an uncommitted update.
-			 */
-			if (F_ISSET(r, WT_REC_UPDATE_RESTORE) &&
-			    *updp != NULL && (uncommitted || prepared)) {
-				r->leave_dirty = true;
-				return (__wt_set_return(session, EBUSY));
-			}
-
-			if (upd->type == WT_UPDATE_BIRTHMARK)
-				skipped_birthmark = true;
-
-			continue;
-		}
-
-		/*
-		 * Lookaside without stable timestamp was taken care of above
-		 * (set to the first uncommitted transaction). Lookaside with
-		 * stable timestamp always takes the first stable update.
-		 */
-		if (*updp == NULL)
-			*updp = upd;
-	}
-
-	/* Keep track of the selected update. */
-	upd = *updp;
-
-	/* Reconciliation should never see an aborted or reserved update. */
-	WT_ASSERT(session, upd == NULL ||
-	    (upd->txnid != WT_TXN_ABORTED && upd->type != WT_UPDATE_RESERVE));
-
-	/* If all of the updates were aborted, quit. */
-	if (first_txn_upd == NULL) {
-		WT_ASSERT(session, upd == NULL);
-		return (0);
-	}
-
-	/* If no updates were skipped, record that we're making progress. */
-	if (upd == first_txn_upd)
-		r->update_used = true;
-
-	/*
-	 * The checkpoint transaction is special.  Make sure we never write
-	 * metadata updates from a checkpoint in a concurrent session.
-	 */
-	WT_ASSERT(session, !WT_IS_METADATA(session->dhandle) ||
-	    upd == NULL || upd->txnid == WT_TXN_NONE ||
-	    upd->txnid != S2C(session)->txn_global.checkpoint_state.id ||
-	    WT_SESSION_IS_CHECKPOINT(session));
-
-	/*
-	 * Track the most recent transaction in the page.  We store this in the
-	 * tree at the end of reconciliation in the service of checkpoints, it
-	 * is used to avoid discarding trees from memory when they have changes
-	 * required to satisfy a snapshot read.
-	 */
-	if (WT_TXNID_LT(r->max_txn, max_txn))
-		r->max_txn = max_txn;
-
-	/* Update the maximum timestamp. */
-	if (first_ts_upd != NULL && r->max_timestamp < first_ts_upd->timestamp)
-		r->max_timestamp = first_ts_upd->timestamp;
-
-	/*
-	 * If the update we chose was a birthmark, or we are doing
-	 * update-restore and we skipped a birthmark, the original on-page
-	 * value must be retained.
-	 */
-	if (upd != NULL &&
-	    (upd->type == WT_UPDATE_BIRTHMARK ||
-	    (F_ISSET(r, WT_REC_UPDATE_RESTORE) && skipped_birthmark)))
-		*updp = NULL;
-
-	/*
-	 * Check if all updates on the page are visible.  If not, it must stay
-	 * dirty unless we are saving updates to the lookaside table.
-	 *
-	 * Updates can be out of transaction ID order (but not out of timestamp
-	 * order), so we track the maximum transaction ID and the newest update
-	 * with a timestamp (if any).
-	 */
-	timestamp = first_ts_upd == NULL ? 0 : first_ts_upd->timestamp;
-	all_visible = upd == first_txn_upd && !(uncommitted || prepared) &&
-	    (F_ISSET(r, WT_REC_VISIBLE_ALL) ?
-	    __wt_txn_visible_all(session, max_txn, timestamp) :
-	    __wt_txn_visible(session, max_txn, timestamp));
-
-	if (all_visible)
-		goto check_original_value;
-
-	r->leave_dirty = true;
-
-	if (F_ISSET(r, WT_REC_VISIBILITY_ERR))
-		WT_PANIC_RET(session, EINVAL,
-		    "reconciliation error, update not visible");
-
-	/*
-	 * If not trying to evict the page, we know what we'll write and we're
-	 * done.
-	 */
-	if (!F_ISSET(r, WT_REC_EVICT))
-		goto check_original_value;
-
-	/*
-	 * We are attempting eviction with changes that are not yet stable
-	 * (i.e. globally visible).  There are two ways to continue, the
-	 * save/restore eviction path or the lookaside table eviction path.
-	 * Both cannot be configured because the paths track different
-	 * information. The update/restore path can handle uncommitted changes,
-	 * by evicting most of the page and then creating a new, smaller page
-	 * to which we re-attach those changes. Lookaside eviction writes
-	 * changes into the lookaside table and restores them on demand if and
-	 * when the page is read back into memory.
-	 *
-	 * Both paths are configured outside of reconciliation: the save/restore
-	 * path is the WT_REC_UPDATE_RESTORE flag, the lookaside table path is
-	 * the WT_REC_LOOKASIDE flag.
-	 */
-	if (!F_ISSET(r, WT_REC_LOOKASIDE | WT_REC_UPDATE_RESTORE))
-		return (__wt_set_return(session, EBUSY));
-	if (uncommitted && !F_ISSET(r, WT_REC_UPDATE_RESTORE))
-		return (__wt_set_return(session, EBUSY));
-
-	WT_ASSERT(session, r->max_txn != WT_TXN_NONE);
-
-	/*
-	 * The order of the updates on the list matters, we can't move only the
-	 * unresolved updates, move the entire update list.
-	 */
-	WT_RET(__rec_update_save(session, r, ins, ripcip, *updp, upd_memsize));
-	upd_saved = true;
-	if (upd_savedp != NULL)
-		*upd_savedp = true;
-
-	/*
-	 * Track the first off-page update when saving history in the lookaside
-	 * table.  When skewing newest, we want the first (non-aborted) update
-	 * after the one stored on the page.  Otherwise, we want the update
-	 * before the on-page update.
-	 */
-	if (F_ISSET(r, WT_REC_LOOKASIDE) && r->las_skew_newest) {
-		if (WT_TXNID_LT(r->unstable_txn, first_upd->txnid))
-			r->unstable_txn = first_upd->txnid;
-		if (first_ts_upd != NULL &&
-		    r->unstable_timestamp < first_ts_upd->timestamp)
-			r->unstable_timestamp = first_ts_upd->timestamp;
-	} else if (F_ISSET(r, WT_REC_LOOKASIDE)) {
-		for (upd = first_upd; upd != *updp; upd = upd->next) {
-			if (upd->txnid == WT_TXN_ABORTED)
-				continue;
-
-			if (upd->txnid != WT_TXN_NONE &&
-			    WT_TXNID_LT(upd->txnid, r->unstable_txn))
-				r->unstable_txn = upd->txnid;
-			if (upd->timestamp < r->unstable_timestamp)
-				r->unstable_timestamp = upd->timestamp;
-		}
-	}
-
-check_original_value:
-	/*
-	 * Paranoia: check that we didn't choose an update that has since been
-	 * rolled back.
-	 */
-	WT_ASSERT(session, *updp == NULL || (*updp)->txnid != WT_TXN_ABORTED);
-
-	/*
-	 * Returning an update means the original on-page value might be lost,
-	 * and that's a problem if there's a reader that needs it.  This call
-	 * makes a copy of the on-page value and if there is a birthmark in the
-	 * update list, replaces it.  We do that any time there are saved
-	 * updates and during reconciliation of a backing overflow record that
-	 * will be physically removed once it's no longer needed.
-	 */
-	if (*updp != NULL && (upd_saved ||
-	    (vpack != NULL && vpack->ovfl &&
-	    vpack->raw != WT_CELL_VALUE_OVFL_RM)))
-		WT_RET(
-		    __rec_append_orig_value(session, page, first_upd, vpack));
-
-	return (0);
-}
-
-/*
- * WT_CHILD_RELEASE, WT_CHILD_RELEASE_ERR --
- *	Macros to clean up during internal-page reconciliation, releasing the
- *	hazard pointer we're holding on child pages.
- */
-#define	WT_CHILD_RELEASE(session, hazard, ref) do {			\
-	if (hazard) {							\
-		(hazard) = false;					\
-		WT_TRET(						\
-		    __wt_page_release(session, ref, WT_READ_NO_EVICT));	\
-	}								\
-} while (0)
-#define	WT_CHILD_RELEASE_ERR(session, hazard, ref) do {			\
-	WT_CHILD_RELEASE(session, hazard, ref);				\
-	WT_ERR(ret);							\
-} while (0)
-
-typedef enum {
-    WT_CHILD_IGNORE,				/* Ignored child */
-    WT_CHILD_MODIFIED,				/* Modified child */
-    WT_CHILD_ORIGINAL,				/* Original child */
-    WT_CHILD_PROXY				/* Deleted child: proxy */
-} WT_CHILD_STATE;
-
-/*
- * __rec_child_deleted --
- *	Handle pages with leaf pages in the WT_REF_DELETED state.
- */
-static int
-__rec_child_deleted(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_REF *ref, WT_CHILD_STATE *statep)
-{
-	WT_PAGE_DELETED *page_del;
-
-	page_del = ref->page_del;
-
-	/*
-	 * Internal pages with child leaf pages in the WT_REF_DELETED state are
-	 * a special case during reconciliation.  First, if the deletion was a
-	 * result of a session truncate call, the deletion may not be visible to
-	 * us. In that case, we proceed as with any change not visible during
-	 * reconciliation by ignoring the change for the purposes of writing the
-	 * internal page.
-	 *
-	 * In this case, there must be an associated page-deleted structure, and
-	 * it holds the transaction ID we care about.
-	 *
-	 * In some cases, there had better not be any updates we can't see.
-	 *
-	 * A visible update to be in READY state (i.e. not in LOCKED or
-	 * PREPARED state), for truly visible to others.
-	 */
-	if (F_ISSET(r, WT_REC_VISIBILITY_ERR) && page_del != NULL &&
-	    __wt_page_del_active(session, ref, false))
-		WT_PANIC_RET(session, EINVAL,
-		    "reconciliation illegally skipped an update");
-
-	/*
-	 * Deal with any underlying disk blocks.
-	 *
-	 * First, check to see if there is an address associated with this leaf:
-	 * if there isn't, we're done, the underlying page is already gone.  If
-	 * the page still exists, check for any transactions in the system that
-	 * might want to see the page's state before it's deleted.
-	 *
-	 * If any such transactions exist, we cannot discard the underlying leaf
-	 * page to the block manager because the transaction may eventually read
-	 * it.  However, this write might be part of a checkpoint, and should we
-	 * recover to that checkpoint, we'll need to delete the leaf page, else
-	 * we'd leak it.  The solution is to write a proxy cell on the internal
-	 * page ensuring the leaf page is eventually discarded.
-	 *
-	 * If no such transactions exist, we can discard the leaf page to the
-	 * block manager and no cell needs to be written at all.  We do this
-	 * outside of the underlying tracking routines because this action is
-	 * permanent and irrevocable.  (Clearing the address means we've lost
-	 * track of the disk address in a permanent way.  This is safe because
-	 * there's no path to reading the leaf page again: if there's ever a
-	 * read into this part of the name space again, the cache read function
-	 * instantiates an entirely new page.)
-	 */
-	if (ref->addr != NULL && !__wt_page_del_active(session, ref, true)) {
-		/*
-		 * Minor memory cleanup: if a truncate call deleted this page
-		 * and we were ever forced to instantiate the page in memory,
-		 * we would have built a list of updates in the page reference
-		 * in order to be able to commit/rollback the truncate. We just
-		 * passed a visibility test, discard the update list.
-		 */
-		if (page_del != NULL) {
-			__wt_free(session, ref->page_del->update_list);
-			__wt_free(session, ref->page_del);
-		}
-
-		WT_RET(__wt_ref_block_free(session, ref));
-	}
-
-	/*
-	 * If the original page is gone, we can skip the slot on the internal
-	 * page.
-	 */
-	if (ref->addr == NULL) {
-		*statep = WT_CHILD_IGNORE;
-		return (0);
-	}
-
-	/*
-	 * Internal pages with deletes that aren't stable cannot be evicted, we
-	 * don't have sufficient information to restore the page's information
-	 * if subsequently read (we wouldn't know which transactions should see
-	 * the original page and which should see the deleted page).
-	 */
-	if (F_ISSET(r, WT_REC_EVICT))
-		return (__wt_set_return(session, EBUSY));
-
-	/*
-	 * If there are deleted child pages we can't discard immediately, keep
-	 * the page dirty so they are eventually freed.
-	 */
-	r->leave_dirty = true;
-
-	/*
-	 * If the original page cannot be freed, we need to keep a slot on the
-	 * page to reference it from the parent page.
-	 *
-	 * If the delete is not visible in this checkpoint, write the original
-	 * address normally. Otherwise, we have to write a proxy record.
-	 * If the delete state is not ready, then delete is not visible as it
-	 * is in prepared state.
-	 */
-	if (!__wt_page_del_active(session, ref, false))
-		*statep = WT_CHILD_PROXY;
-
-	return (0);
-}
-
-/*
- * __rec_child_modify --
- *	Return if the internal page's child references any modifications.
- */
-static int
-__rec_child_modify(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_REF *ref, bool *hazardp, WT_CHILD_STATE *statep)
-{
-	WT_DECL_RET;
-	WT_PAGE_MODIFY *mod;
-
-	/* We may acquire a hazard pointer our caller must release. */
-	*hazardp = false;
-
-	/* Default to using the original child address. */
-	*statep = WT_CHILD_ORIGINAL;
-
-	/*
-	 * This function is called when walking an internal page to decide how
-	 * to handle child pages referenced by the internal page.
-	 *
-	 * Internal pages are reconciled for two reasons: first, when evicting
-	 * an internal page, second by the checkpoint code when writing internal
-	 * pages.  During eviction, all pages should be in the WT_REF_DISK or
-	 * WT_REF_DELETED state. During checkpoint, eviction that might affect
-	 * review of an internal page is prohibited, however, as the subtree is
-	 * not reserved for our exclusive use, there are other page states that
-	 * must be considered.
-	 */
-	for (;; __wt_yield()) {
-		switch (r->tested_ref_state = ref->state) {
-		case WT_REF_DISK:
-			/* On disk, not modified by definition. */
-			goto done;
-
-		case WT_REF_DELETED:
-			/*
-			 * The child is in a deleted state.
-			 *
-			 * It's possible the state could change underneath us as
-			 * the page is read in, and we can race between checking
-			 * for a deleted state and looking at the transaction ID
-			 * to see if the delete is visible to us.  Lock down the
-			 * structure.
-			 */
-			if (!WT_REF_CAS_STATE(
-			    session, ref, WT_REF_DELETED, WT_REF_LOCKED))
-				break;
-			ret = __rec_child_deleted(session, r, ref, statep);
-			WT_REF_SET_STATE(ref, WT_REF_DELETED);
-			goto done;
-
-		case WT_REF_LOCKED:
-			/*
-			 * Locked.
-			 *
-			 * We should never be here during eviction, active child
-			 * pages in an evicted page's subtree fails the eviction
-			 * attempt.
-			 */
-			WT_ASSERT(session, !F_ISSET(r, WT_REC_EVICT));
-			if (F_ISSET(r, WT_REC_EVICT))
-				return (__wt_set_return(session, EBUSY));
-
-			/*
-			 * If called during checkpoint, the child is being
-			 * considered by the eviction server or the child is a
-			 * truncated page being read.  The eviction may have
-			 * started before the checkpoint and so we must wait
-			 * for the eviction to be resolved.  I suspect we could
-			 * handle reads of truncated pages, but we can't
-			 * distinguish between the two and reads of truncated
-			 * pages aren't expected to be common.
-			 */
-			break;
-
-		case WT_REF_LIMBO:
-			WT_ASSERT(session, !F_ISSET(r, WT_REC_EVICT));
-			/* FALLTHROUGH */
-		case WT_REF_LOOKASIDE:
-			/*
-			 * On disk or in cache with lookaside updates.
-			 *
-			 * We should never be here during eviction: active
-			 * child pages in an evicted page's subtree fails the
-			 * eviction attempt.
-			 */
-			if (F_ISSET(r, WT_REC_EVICT) &&
-			    __wt_page_las_active(session, ref)) {
-				WT_ASSERT(session, false);
-				return (__wt_set_return(session, EBUSY));
-			}
-
-			/*
-			 * A page evicted with lookaside entries may not have
-			 * an address, if no updates were visible to
-			 * reconciliation.  Any child pages in that state
-			 * should be ignored.
-			 */
-			if (ref->addr == NULL) {
-				*statep = WT_CHILD_IGNORE;
-				WT_CHILD_RELEASE(session, *hazardp, ref);
-			}
-			goto done;
-
-		case WT_REF_MEM:
-			/*
-			 * In memory.
-			 *
-			 * We should never be here during eviction, active child
-			 * pages in an evicted page's subtree fails the eviction
-			 * attempt.
-			 */
-			WT_ASSERT(session, !F_ISSET(r, WT_REC_EVICT));
-			if (F_ISSET(r, WT_REC_EVICT))
-				return (__wt_set_return(session, EBUSY));
-
-			/*
-			 * If called during checkpoint, acquire a hazard pointer
-			 * so the child isn't evicted, it's an in-memory case.
-			 *
-			 * This call cannot return split/restart, we have a lock
-			 * on the parent which prevents a child page split.
-			 *
-			 * Set WT_READ_NO_WAIT because we're only interested in
-			 * the WT_REF's final state. Pages in transition might
-			 * change WT_REF state during our read, and then return
-			 * WT_NOTFOUND to us. In that case, loop and look again.
-			 */
-			ret = __wt_page_in(session, ref,
-			    WT_READ_CACHE | WT_READ_NO_EVICT |
-			    WT_READ_NO_GEN | WT_READ_NO_WAIT);
-			if (ret == WT_NOTFOUND) {
-				ret = 0;
-				break;
-			}
-			WT_RET(ret);
-			*hazardp = true;
-			goto in_memory;
-
-		case WT_REF_READING:
-			/*
-			 * Being read, not modified by definition.
-			 *
-			 * We should never be here during eviction, active child
-			 * pages in an evicted page's subtree fails the eviction
-			 * attempt.
-			 */
-			WT_ASSERT(session, !F_ISSET(r, WT_REC_EVICT));
-			if (F_ISSET(r, WT_REC_EVICT))
-				return (__wt_set_return(session, EBUSY));
-			goto done;
-
-		case WT_REF_SPLIT:
-			/*
-			 * The page was split out from under us.
-			 *
-			 * We should never be here during eviction, active child
-			 * pages in an evicted page's subtree fails the eviction
-			 * attempt.
-			 *
-			 * We should never be here during checkpoint, dirty page
-			 * eviction is shutout during checkpoint, all splits in
-			 * process will have completed before we walk any pages
-			 * for checkpoint.
-			 */
-			WT_ASSERT(session, WT_REF_SPLIT != WT_REF_SPLIT);
-			return (__wt_set_return(session, EBUSY));
-
-		WT_ILLEGAL_VALUE(session, r->tested_ref_state);
-		}
-		WT_STAT_CONN_INCR(session, child_modify_blocked_page);
-	}
-
-in_memory:
-	/*
-	 * In-memory states: the child is potentially modified if the page's
-	 * modify structure has been instantiated. If the modify structure
-	 * exists and the page has actually been modified, set that state.
-	 * If that's not the case, we would normally use the original cell's
-	 * disk address as our reference, however there are two special cases,
-	 * both flagged by a missing block address.
-	 *
-	 * First, if forced to instantiate a deleted child page and it's never
-	 * modified, we end up here with a page that has a modify structure, no
-	 * modifications, and no disk address. Ignore those pages, they're not
-	 * modified and there is no reason to write the cell.
-	 *
-	 * Second, insert splits are permitted during checkpoint. When doing the
-	 * final checkpoint pass, we first walk the internal page's page-index
-	 * and write out any dirty pages we find, then we write out the internal
-	 * page in post-order traversal. If we found the split page in the first
-	 * step, it will have an address; if we didn't find the split page in
-	 * the first step, it won't have an address and we ignore it, it's not
-	 * part of the checkpoint.
-	 */
-	mod = ref->page->modify;
-	if (mod != NULL && mod->rec_result != 0)
-		*statep = WT_CHILD_MODIFIED;
-	else if (ref->addr == NULL) {
-		*statep = WT_CHILD_IGNORE;
-		WT_CHILD_RELEASE(session, *hazardp, ref);
-	}
-
-done:	WT_DIAGNOSTIC_YIELD;
-	return (ret);
-}
-
-/*
- * __rec_incr --
- *	Update the memory tracking structure for a set of new entries.
- */
-static inline void
-__rec_incr(WT_SESSION_IMPL *session, WT_RECONCILE *r, uint32_t v, size_t size)
-{
-	/*
-	 * The buffer code is fragile and prone to off-by-one errors -- check
-	 * for overflow in diagnostic mode.
-	 */
-	WT_ASSERT(session, r->space_avail >= size);
-	WT_ASSERT(session, WT_BLOCK_FITS(r->first_free, size,
-	    r->cur_ptr->image.mem, r->cur_ptr->image.memsize));
-
-	r->entries += v;
-	r->space_avail -= size;
-	r->first_free += size;
-
-	/*
-	 * If offset for the minimum split size boundary is not set, we have not
-	 * yet reached the minimum boundary, reduce the space available for it.
-	 */
-	if (r->cur_ptr->min_offset == 0) {
-		if (r->min_space_avail >= size)
-			r->min_space_avail -= size;
-		else
-			r->min_space_avail = 0;
-	}
-}
-
-/*
- * __rec_copy_incr --
- *	Copy a key/value cell and buffer pair into the new image.
- */
-static inline void
-__rec_copy_incr(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_KV *kv)
-{
-	size_t len;
-	uint8_t *p, *t;
-
-	/*
-	 * If there's only one chunk of data to copy (because the cell and data
-	 * are being copied from the original disk page), the cell length won't
-	 * be set, the WT_ITEM data/length will reference the data to be copied.
-	 *
-	 * WT_CELLs are typically small, 1 or 2 bytes -- don't call memcpy, do
-	 * the copy in-line.
-	 */
-	for (p = r->first_free,
-	    t = (uint8_t *)&kv->cell, len = kv->cell_len; len > 0; --len)
-		*p++ = *t++;
-
-	/* The data can be quite large -- call memcpy. */
-	if (kv->buf.size != 0)
-		memcpy(p, kv->buf.data, kv->buf.size);
-
-	WT_ASSERT(session, kv->len == kv->cell_len + kv->buf.size);
-	__rec_incr(session, r, 1, kv->len);
-}
-
-/*
- * __rec_dict_replace --
- *	Check for a dictionary match.
- */
-static int
-__rec_dict_replace(
-    WT_SESSION_IMPL *session, WT_RECONCILE *r, uint64_t rle, WT_KV *val)
-{
-	WT_DICTIONARY *dp;
-	uint64_t offset;
-
-	/*
-	 * We optionally create a dictionary of values and only write a unique
-	 * value once per page, using a special "copy" cell for all subsequent
-	 * copies of the value.  We have to do the cell build and resolution at
-	 * this low level because we need physical cell offsets for the page.
-	 *
-	 * Sanity check: short-data cells can be smaller than dictionary-copy
-	 * cells.  If the data is already small, don't bother doing the work.
-	 * This isn't just work avoidance: on-page cells can't grow as a result
-	 * of writing a dictionary-copy cell, the reconciliation functions do a
-	 * split-boundary test based on the size required by the value's cell;
-	 * if we grow the cell after that test we'll potentially write off the
-	 * end of the buffer's memory.
-	 */
-	if (val->buf.size <= WT_INTPACK32_MAXSIZE)
-		return (0);
-	WT_RET(__rec_dictionary_lookup(session, r, val, &dp));
-	if (dp == NULL)
-		return (0);
-
-	/*
-	 * If the dictionary offset isn't set, we're creating a new entry in the
-	 * dictionary, set its location.
-	 *
-	 * If the dictionary offset is set, we have a matching value. Create a
-	 * copy cell instead.
-	 */
-	if (dp->offset == 0)
-		dp->offset = WT_PTRDIFF32(r->first_free, r->cur_ptr->image.mem);
-	else {
-		/*
-		 * The offset is the byte offset from this cell to the previous,
-		 * matching cell, NOT the byte offset from the beginning of the
-		 * page.
-		 */
-		offset = (uint64_t)WT_PTRDIFF(r->first_free,
-		    (uint8_t *)r->cur_ptr->image.mem + dp->offset);
-		val->len = val->cell_len =
-		    __wt_cell_pack_copy(&val->cell, rle, offset);
-		val->buf.data = NULL;
-		val->buf.size = 0;
-	}
-	return (0);
-}
-
-/*
- * __rec_key_state_update --
- *	Update prefix and suffix compression based on the last key.
- */
-static inline void
-__rec_key_state_update(WT_RECONCILE *r, bool ovfl_key)
-{
-	WT_ITEM *a;
-
-	/*
-	 * If writing an overflow key onto the page, don't update the "last key"
-	 * value, and leave the state of prefix compression alone.  (If we are
-	 * currently doing prefix compression, we have a key state which will
-	 * continue to work, we're just skipping the key just created because
-	 * it's an overflow key and doesn't participate in prefix compression.
-	 * If we are not currently doing prefix compression, we can't start, an
-	 * overflow key doesn't give us any state.)
-	 *
-	 * Additionally, if we wrote an overflow key onto the page, turn off the
-	 * suffix compression of row-store internal node keys.  (When we split,
-	 * "last key" is the largest key on the previous page, and "cur key" is
-	 * the first key on the next page, which is being promoted.  In some
-	 * cases we can discard bytes from the "cur key" that are not needed to
-	 * distinguish between the "last key" and "cur key", compressing the
-	 * size of keys on internal nodes.  If we just built an overflow key,
-	 * we're not going to update the "last key", making suffix compression
-	 * impossible for the next key. Alternatively, we could remember where
-	 * the last key was on the page, detect it's an overflow key, read it
-	 * from disk and do suffix compression, but that's too much work for an
-	 * unlikely event.)
-	 *
-	 * If we're not writing an overflow key on the page, update the last-key
-	 * value and turn on both prefix and suffix compression.
-	 */
-	if (ovfl_key)
-		r->key_sfx_compress = false;
-	else {
-		a = r->cur;
-		r->cur = r->last;
-		r->last = a;
-
-		r->key_pfx_compress = r->key_pfx_compress_conf;
-		r->key_sfx_compress = r->key_sfx_compress_conf;
-	}
-}
-
-/*
- * Macros from fixed-length entries to/from bytes.
- */
-#define	WT_FIX_BYTES_TO_ENTRIES(btree, bytes)				\
-    ((uint32_t)((((bytes) * 8) / (btree)->bitcnt)))
-#define	WT_FIX_ENTRIES_TO_BYTES(btree, entries)				\
-	((uint32_t)WT_ALIGN((entries) * (btree)->bitcnt, 8))
 
 /*
  * __rec_leaf_page_max --
@@ -2057,35 +854,6 @@ __rec_leaf_page_max(WT_SESSION_IMPL *session, WT_RECONCILE *r)
 }
 
 /*
- * __rec_need_split --
- *	Check whether adding some bytes to the page requires a split.
- */
-static bool
-__rec_need_split(WT_RECONCILE *r, size_t len)
-{
-	/*
-	 * In the case of a row-store leaf page, trigger a split if a threshold
-	 * number of saved updates is reached. This allows pages to split for
-	 * update/restore and lookaside eviction when there is no visible data
-	 * causing the disk image to grow.
-	 *
-	 * In the case of small pages or large keys, we might try to split when
-	 * a page has no updates or entries, which isn't possible. To consider
-	 * update/restore or lookaside information, require either page entries
-	 * or updates that will be attached to the image. The limit is one of
-	 * either, but it doesn't make sense to create pages or images with few
-	 * entries or updates, even where page sizes are small (especially as
-	 * updates that will eventually become overflow items can throw off our
-	 * calculations). Bound the combination at something reasonable.
-	 */
-	if (r->page->type == WT_PAGE_ROW_LEAF && r->entries + r->supd_next > 10)
-		len += r->supd_memsize;
-
-	/* Check for the disk image crossing a boundary. */
-	return (WT_CHECK_CROSSING_BND(r, len));
-}
-
-/*
  * __wt_split_page_size --
  *	Given a split percentage, calculate split page size in bytes.
  */
@@ -2123,8 +891,8 @@ __wt_split_page_size(int split_pct, uint32_t maxpagesize, uint32_t allocsize)
  *	Initialize a single chunk structure.
  */
 static int
-__rec_split_chunk_init(
-    WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_CHUNK *chunk, size_t memsize)
+__rec_split_chunk_init(WT_SESSION_IMPL *session,
+    WT_RECONCILE *r, WT_REC_CHUNK *chunk, size_t memsize)
 {
 	chunk->min_recno = WT_RECNO_OOB;
 	chunk->min_entries = 0;
@@ -2158,16 +926,16 @@ __rec_split_chunk_init(
 }
 
 /*
- * __rec_split_init --
+ * __wt_rec_split_init --
  *	Initialization for the reconciliation split functions.
  */
-static int
-__rec_split_init(WT_SESSION_IMPL *session,
+int
+__wt_rec_split_init(WT_SESSION_IMPL *session,
     WT_RECONCILE *r, WT_PAGE *page, uint64_t recno, uint64_t max)
 {
 	WT_BM *bm;
 	WT_BTREE *btree;
-	WT_CHUNK *chunk;
+	WT_REC_CHUNK *chunk;
 	WT_REF *ref;
 	size_t corrected_page_size, disk_img_buf_size;
 
@@ -2463,21 +1231,21 @@ __rec_split_grow(WT_SESSION_IMPL *session, WT_RECONCILE *r, size_t add_len)
 }
 
 /*
- * __rec_split --
+ * __wt_rec_split --
  *	Handle the page reconciliation bookkeeping.  (Did you know "bookkeeper"
  *	has 3 doubled letters in a row?  Sweet-tooth does, too.)
  */
-static int
-__rec_split(WT_SESSION_IMPL *session, WT_RECONCILE *r, size_t next_len)
+int
+__wt_rec_split(WT_SESSION_IMPL *session, WT_RECONCILE *r, size_t next_len)
 {
 	WT_BTREE *btree;
-	WT_CHUNK *tmp;
+	WT_REC_CHUNK *tmp;
 	size_t inuse;
 
 	btree = S2BT(session);
 
 	/* Fixed length col store can call with next_len 0 */
-	WT_ASSERT(session, next_len == 0 || __rec_need_split(r, next_len));
+	WT_ASSERT(session, next_len == 0 || __wt_rec_need_split(r, next_len));
 
 	/*
 	 * We should never split during salvage, and we're about to drop core
@@ -2495,11 +1263,11 @@ __rec_split(WT_SESSION_IMPL *session, WT_RECONCILE *r, size_t next_len)
 	 * Additionally, grow the buffer to contain the current item if we
 	 * haven't already consumed a reasonable portion of a split chunk.
 	 */
-	if (inuse < r->split_size / 2 && !__rec_need_split(r, 0))
+	if (inuse < r->split_size / 2 && !__wt_rec_need_split(r, 0))
 		goto done;
 
 	/* All page boundaries reset the dictionary. */
-	__rec_dictionary_reset(r);
+	__wt_rec_dictionary_reset(r);
 
 	/* Set the number of entries and size for the just finished chunk. */
 	r->cur_ptr->entries = r->entries;
@@ -2567,18 +1335,18 @@ done:  	/*
 }
 
 /*
- * __rec_split_crossing_bnd --
+ * __wt_rec_split_crossing_bnd --
  * 	Save the details for the minimum split size boundary or call for a
  * 	split.
  */
-static inline int
-__rec_split_crossing_bnd(
+int
+__wt_rec_split_crossing_bnd(
     WT_SESSION_IMPL *session, WT_RECONCILE *r, size_t next_len)
 {
 	WT_BTREE *btree;
 	size_t min_offset;
 
-	WT_ASSERT(session, __rec_need_split(r, next_len));
+	WT_ASSERT(session, __wt_rec_need_split(r, next_len));
 
 	/*
 	 * If crossing the minimum split size boundary, store the boundary
@@ -2587,7 +1355,7 @@ __rec_split_crossing_bnd(
 	 * large enough, just split at this point.
 	 */
 	if (WT_CROSSING_MIN_BND(r, next_len) &&
-	    !WT_CROSSING_SPLIT_BND(r, next_len) && !__rec_need_split(r, 0)) {
+	    !WT_CROSSING_SPLIT_BND(r, next_len) && !__wt_rec_need_split(r, 0)) {
 		btree = S2BT(session);
 		WT_ASSERT(session, r->cur_ptr->min_offset == 0);
 
@@ -2609,13 +1377,13 @@ __rec_split_crossing_bnd(
 			    session, r, &r->cur_ptr->min_key, r->page->type));
 
 		/* All page boundaries reset the dictionary. */
-		__rec_dictionary_reset(r);
+		__wt_rec_dictionary_reset(r);
 
 		return (0);
 	}
 
 	/* We are crossing a split boundary */
-	return (__rec_split(session, r, next_len));
+	return (__wt_rec_split(session, r, next_len));
 }
 
 /*
@@ -2632,8 +1400,8 @@ static int
 __rec_split_finish_process_prev(WT_SESSION_IMPL *session, WT_RECONCILE *r)
 {
 	WT_BTREE *btree;
-	WT_CHUNK *cur_ptr, *prev_ptr, *tmp;
 	WT_PAGE_HEADER *dsk;
+	WT_REC_CHUNK *cur_ptr, *prev_ptr, *tmp;
 	size_t combined_size, len_to_move;
 	uint8_t *cur_dsk_start;
 
@@ -2714,11 +1482,11 @@ __rec_split_finish_process_prev(WT_SESSION_IMPL *session, WT_RECONCILE *r)
 }
 
 /*
- * __rec_split_finish --
+ * __wt_rec_split_finish --
  *	Finish processing a page.
  */
-static int
-__rec_split_finish(WT_SESSION_IMPL *session, WT_RECONCILE *r)
+int
+__wt_rec_split_finish(WT_SESSION_IMPL *session, WT_RECONCILE *r)
 {
 	/*
 	 * We're done reconciling, write the final page. We may arrive here with
@@ -2771,13 +1539,13 @@ __rec_supd_move(
  */
 static int
 __rec_split_write_supd(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_CHUNK *chunk, WT_MULTI *multi, bool last_block)
+    WT_RECONCILE *r, WT_REC_CHUNK *chunk, WT_MULTI *multi, bool last_block)
 {
 	WT_BTREE *btree;
-	WT_CHUNK *next;
 	WT_DECL_ITEM(key);
 	WT_DECL_RET;
 	WT_PAGE *page;
+	WT_REC_CHUNK *next;
 	WT_SAVE_UPD *supd;
 	WT_UPDATE *upd;
 	uint32_t i, j;
@@ -2876,7 +1644,7 @@ err:	__wt_scr_free(session, &key);
  */
 static void
 __rec_split_write_header(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_CHUNK *chunk, WT_MULTI *multi, WT_PAGE_HEADER *dsk)
+    WT_RECONCILE *r, WT_REC_CHUNK *chunk, WT_MULTI *multi, WT_PAGE_HEADER *dsk)
 {
 	WT_BTREE *btree;
 	WT_PAGE *page;
@@ -3088,7 +1856,7 @@ __rec_compression_adjust(WT_SESSION_IMPL *session,
  */
 static int
 __rec_split_write(WT_SESSION_IMPL *session, WT_RECONCILE *r,
-    WT_CHUNK *chunk, WT_ITEM *compressed_image, bool last_block)
+    WT_REC_CHUNK *chunk, WT_ITEM *compressed_image, bool last_block)
 {
 	WT_BTREE *btree;
 	WT_MULTI *multi;
@@ -3304,7 +2072,7 @@ __wt_bulk_init(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
 
 	recno = btree->type == BTREE_ROW ? WT_RECNO_OOB : 1;
 
-	return (__rec_split_init(session,
+	return (__wt_rec_split_init(session,
 	    r, cbulk->leaf, recno, btree->maxleafpage_precomp));
 }
 
@@ -3326,7 +2094,7 @@ __wt_bulk_wrapup(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
 	switch (btree->type) {
 	case BTREE_COL_FIX:
 		if (cbulk->entry != 0)
-			__rec_incr(session, r, cbulk->entry,
+			__wt_rec_incr(session, r, cbulk->entry,
 			    __bitstr_size(
 			    (size_t)cbulk->entry * btree->bitcnt));
 		break;
@@ -3338,7 +2106,7 @@ __wt_bulk_wrapup(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
 		break;
 	}
 
-	WT_RET(__rec_split_finish(session, r));
+	WT_RET(__wt_rec_split_finish(session, r));
 	WT_RET(__rec_write_wrapup(session, r, r->page));
 	__rec_write_page_status(session, r);
 
@@ -3349,1912 +2117,6 @@ __wt_bulk_wrapup(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
 
 	__rec_cleanup(session, r);
 	__rec_destroy(session, &cbulk->reconcile);
-
-	return (0);
-}
-
-/*
- * __wt_bulk_insert_row --
- *	Row-store bulk insert.
- */
-int
-__wt_bulk_insert_row(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
-{
-	WT_BTREE *btree;
-	WT_CURSOR *cursor;
-	WT_KV *key, *val;
-	WT_RECONCILE *r;
-	bool ovfl_key;
-
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-	cursor = &cbulk->cbt.iface;
-
-	key = &r->k;
-	val = &r->v;
-	WT_RET(__rec_cell_build_leaf_key(session, r,	/* Build key cell */
-	    cursor->key.data, cursor->key.size, &ovfl_key));
-	WT_RET(__rec_cell_build_val(session, r,		/* Build value cell */
-	    cursor->value.data, cursor->value.size, (uint64_t)0));
-
-	/* Boundary: split or write the page. */
-	if (WT_CROSSING_SPLIT_BND(r, key->len + val->len)) {
-		/*
-		 * Turn off prefix compression until a full key written to the
-		 * new page, and (unless already working with an overflow key),
-		 * rebuild the key without compression.
-		 */
-		if (r->key_pfx_compress_conf) {
-			r->key_pfx_compress = false;
-			if (!ovfl_key)
-				WT_RET(__rec_cell_build_leaf_key(
-				    session, r, NULL, 0, &ovfl_key));
-		}
-		WT_RET(__rec_split_crossing_bnd(
-		    session, r, key->len + val->len));
-	}
-
-	/* Copy the key/value pair onto the page. */
-	__rec_copy_incr(session, r, key);
-	if (val->len == 0)
-		r->any_empty_value = true;
-	else {
-		r->all_empty_value = false;
-		if (btree->dictionary)
-			WT_RET(__rec_dict_replace(session, r, 0, val));
-		__rec_copy_incr(session, r, val);
-	}
-
-	/* Update compression state. */
-	__rec_key_state_update(r, ovfl_key);
-
-	return (0);
-}
-
-/*
- * __rec_col_fix_bulk_insert_split_check --
- *	Check if a bulk-loaded fixed-length column store page needs to split.
- */
-static inline int
-__rec_col_fix_bulk_insert_split_check(WT_CURSOR_BULK *cbulk)
-{
-	WT_BTREE *btree;
-	WT_RECONCILE *r;
-	WT_SESSION_IMPL *session;
-
-	session = (WT_SESSION_IMPL *)cbulk->cbt.iface.session;
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-
-	if (cbulk->entry == cbulk->nrecs) {
-		if (cbulk->entry != 0) {
-			/*
-			 * If everything didn't fit, update the counters and
-			 * split.
-			 *
-			 * Boundary: split or write the page.
-			 *
-			 * No need to have a minimum split size boundary, all
-			 * pages are filled 100% except the last, allowing it to
-			 * grow in the future.
-			 */
-			__rec_incr(session, r, cbulk->entry,
-			    __bitstr_size(
-			    (size_t)cbulk->entry * btree->bitcnt));
-			WT_RET(__rec_split(session, r, 0));
-		}
-		cbulk->entry = 0;
-		cbulk->nrecs = WT_FIX_BYTES_TO_ENTRIES(btree, r->space_avail);
-	}
-	return (0);
-}
-
-/*
- * __wt_bulk_insert_fix --
- *	Fixed-length column-store bulk insert.
- */
-int
-__wt_bulk_insert_fix(
-    WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk, bool deleted)
-{
-	WT_BTREE *btree;
-	WT_CURSOR *cursor;
-	WT_RECONCILE *r;
-
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-	cursor = &cbulk->cbt.iface;
-
-	WT_RET(__rec_col_fix_bulk_insert_split_check(cbulk));
-	__bit_setv(r->first_free, cbulk->entry,
-	    btree->bitcnt, deleted ? 0 : ((uint8_t *)cursor->value.data)[0]);
-	++cbulk->entry;
-	++r->recno;
-
-	return (0);
-}
-
-/*
- * __wt_bulk_insert_fix_bitmap --
- *	Fixed-length column-store bulk insert.
- */
-int
-__wt_bulk_insert_fix_bitmap(WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk)
-{
-	WT_BTREE *btree;
-	WT_CURSOR *cursor;
-	WT_RECONCILE *r;
-	uint32_t entries, offset, page_entries, page_size;
-	const uint8_t *data;
-
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-	cursor = &cbulk->cbt.iface;
-
-	if (((r->recno - 1) * btree->bitcnt) & 0x7)
-		WT_RET_MSG(session, EINVAL,
-		    "Bulk bitmap load not aligned on a byte boundary");
-	for (data = cursor->value.data,
-	    entries = (uint32_t)cursor->value.size;
-	    entries > 0;
-	    entries -= page_entries, data += page_size) {
-		WT_RET(__rec_col_fix_bulk_insert_split_check(cbulk));
-
-		page_entries = WT_MIN(entries, cbulk->nrecs - cbulk->entry);
-		page_size = __bitstr_size(page_entries * btree->bitcnt);
-		offset = __bitstr_size(cbulk->entry * btree->bitcnt);
-		memcpy(r->first_free + offset, data, page_size);
-		cbulk->entry += page_entries;
-		r->recno += page_entries;
-	}
-	return (0);
-}
-
-/*
- * __wt_bulk_insert_var --
- *	Variable-length column-store bulk insert.
- */
-int
-__wt_bulk_insert_var(
-    WT_SESSION_IMPL *session, WT_CURSOR_BULK *cbulk, bool deleted)
-{
-	WT_BTREE *btree;
-	WT_KV *val;
-	WT_RECONCILE *r;
-
-	r = cbulk->reconcile;
-	btree = S2BT(session);
-
-	val = &r->v;
-	if (deleted) {
-		val->cell_len = __wt_cell_pack_del(&val->cell, cbulk->rle);
-		val->buf.data = NULL;
-		val->buf.size = 0;
-		val->len = val->cell_len;
-	} else
-		/*
-		 * Store the bulk cursor's last buffer, not the current value,
-		 * we're tracking duplicates, which means we want the previous
-		 * value seen, not the current value.
-		 */
-		WT_RET(__rec_cell_build_val(session,
-		    r, cbulk->last.data, cbulk->last.size, cbulk->rle));
-
-	/* Boundary: split or write the page. */
-	if (WT_CROSSING_SPLIT_BND(r, val->len))
-		WT_RET(__rec_split_crossing_bnd(session, r, val->len));
-
-	/* Copy the value onto the page. */
-	if (btree->dictionary)
-		WT_RET(__rec_dict_replace(session, r, cbulk->rle, val));
-	__rec_copy_incr(session, r, val);
-
-	/* Update the starting record number in case we split. */
-	r->recno += cbulk->rle;
-
-	return (0);
-}
-
-/*
- * __rec_vtype --
- *	Return a value cell's address type.
- */
-static inline u_int
-__rec_vtype(WT_ADDR *addr)
-{
-	if (addr->type == WT_ADDR_INT)
-		return (WT_CELL_ADDR_INT);
-	if (addr->type == WT_ADDR_LEAF)
-		return (WT_CELL_ADDR_LEAF);
-	return (WT_CELL_ADDR_LEAF_NO);
-}
-
-/*
- * __rec_col_int --
- *	Reconcile a column-store internal page.
- */
-static int
-__rec_col_int(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_REF *pageref)
-{
-	WT_ADDR *addr;
-	WT_BTREE *btree;
-	WT_CELL_UNPACK *vpack, _vpack;
-	WT_CHILD_STATE state;
-	WT_DECL_RET;
-	WT_KV *val;
-	WT_PAGE *child, *page;
-	WT_REF *ref;
-	bool hazard;
-
-	btree = S2BT(session);
-	page = pageref->page;
-	child = NULL;
-	hazard = false;
-
-	val = &r->v;
-	vpack = &_vpack;
-
-	WT_RET(__rec_split_init(session,
-	    r, page, pageref->ref_recno, btree->maxintlpage_precomp));
-
-	/* For each entry in the in-memory page... */
-	WT_INTL_FOREACH_BEGIN(session, page, ref) {
-		/* Update the starting record number in case we split. */
-		r->recno = ref->ref_recno;
-
-		/*
-		 * Modified child.
-		 * The page may be emptied or internally created during a split.
-		 * Deleted/split pages are merged into the parent and discarded.
-		 */
-		WT_ERR(__rec_child_modify(session, r, ref, &hazard, &state));
-		addr = NULL;
-		child = ref->page;
-
-		switch (state) {
-		case WT_CHILD_IGNORE:
-			/* Ignored child. */
-			WT_CHILD_RELEASE_ERR(session, hazard, ref);
-			continue;
-
-		case WT_CHILD_MODIFIED:
-			/*
-			 * Modified child. Empty pages are merged into the
-			 * parent and discarded.
-			 */
-			switch (child->modify->rec_result) {
-			case WT_PM_REC_EMPTY:
-				/*
-				 * Column-store pages are almost never empty, as
-				 * discarding a page would remove a chunk of the
-				 * name space.  The exceptions are pages created
-				 * when the tree is created, and never filled.
-				 */
-				WT_CHILD_RELEASE_ERR(session, hazard, ref);
-				continue;
-			case WT_PM_REC_MULTIBLOCK:
-				WT_ERR(__rec_col_merge(session, r, child));
-				WT_CHILD_RELEASE_ERR(session, hazard, ref);
-				continue;
-			case WT_PM_REC_REPLACE:
-				addr = &child->modify->mod_replace;
-				break;
-			WT_ILLEGAL_VALUE_ERR(
-			    session, child->modify->rec_result);
-			}
-			break;
-		case WT_CHILD_ORIGINAL:
-			/* Original child. */
-			break;
-		case WT_CHILD_PROXY:
-			/*
-			 * Deleted child where we write a proxy cell, not yet
-			 * supported for column-store.
-			 */
-			WT_ERR(__wt_illegal_value(session, state));
-		}
-
-		/*
-		 * Build the value cell.  The child page address is in one of 3
-		 * places: if the page was replaced, the page's modify structure
-		 * references it and we built the value cell just above in the
-		 * switch statement.  Else, the WT_REF->addr reference points to
-		 * an on-page cell or an off-page WT_ADDR structure: if it's an
-		 * on-page cell and we copy it from the page, else build a new
-		 * cell.
-		 */
-		if (addr == NULL && __wt_off_page(page, ref->addr))
-			addr = ref->addr;
-		if (addr == NULL) {
-			__wt_cell_unpack(ref->addr, vpack);
-			val->buf.data = ref->addr;
-			val->buf.size = __wt_cell_total_len(vpack);
-			val->cell_len = 0;
-			val->len = val->buf.size;
-		} else
-			__rec_cell_build_addr(session, r,
-			    addr->addr, addr->size,
-			    __rec_vtype(addr), ref->ref_recno);
-		WT_CHILD_RELEASE_ERR(session, hazard, ref);
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, val->len))
-			WT_ERR(__rec_split_crossing_bnd(session, r, val->len));
-
-		/* Copy the value onto the page. */
-		__rec_copy_incr(session, r, val);
-	} WT_INTL_FOREACH_END;
-
-	/* Write the remnant page. */
-	return (__rec_split_finish(session, r));
-
-err:	WT_CHILD_RELEASE(session, hazard, ref);
-	return (ret);
-}
-
-/*
- * __rec_col_merge --
- *	Merge in a split page.
- */
-static int
-__rec_col_merge(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_PAGE *page)
-{
-	WT_ADDR *addr;
-	WT_KV *val;
-	WT_MULTI *multi;
-	WT_PAGE_MODIFY *mod;
-	uint32_t i;
-
-	mod = page->modify;
-
-	val = &r->v;
-
-	/* For each entry in the split array... */
-	for (multi = mod->mod_multi,
-	    i = 0; i < mod->mod_multi_entries; ++multi, ++i) {
-		/* Update the starting record number in case we split. */
-		r->recno = multi->key.recno;
-
-		/* Build the value cell. */
-		addr = &multi->addr;
-		__rec_cell_build_addr(session, r,
-		    addr->addr, addr->size, __rec_vtype(addr), r->recno);
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, val->len))
-			WT_RET(__rec_split_crossing_bnd(session, r, val->len));
-
-		/* Copy the value onto the page. */
-		__rec_copy_incr(session, r, val);
-	}
-	return (0);
-}
-
-/*
- * __rec_col_fix --
- *	Reconcile a fixed-width, column-store leaf page.
- */
-static int
-__rec_col_fix(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_REF *pageref)
-{
-	WT_BTREE *btree;
-	WT_INSERT *ins;
-	WT_PAGE *page;
-	WT_UPDATE *upd;
-	uint64_t recno;
-	uint32_t entry, nrecs;
-
-	btree = S2BT(session);
-	page = pageref->page;
-
-	WT_RET(__rec_split_init(
-	    session, r, page, pageref->ref_recno, btree->maxleafpage));
-
-	/* Copy the original, disk-image bytes into place. */
-	memcpy(r->first_free, page->pg_fix_bitf,
-	    __bitstr_size((size_t)page->entries * btree->bitcnt));
-
-	/* Update any changes to the original on-page data items. */
-	WT_SKIP_FOREACH(ins, WT_COL_UPDATE_SINGLE(page)) {
-		WT_RET(__rec_txn_read(session, r, ins, NULL, NULL, NULL, &upd));
-		if (upd != NULL)
-			__bit_setv(r->first_free,
-			    WT_INSERT_RECNO(ins) - pageref->ref_recno,
-			    btree->bitcnt, *upd->data);
-	}
-
-	/* Calculate the number of entries per page remainder. */
-	entry = page->entries;
-	nrecs = WT_FIX_BYTES_TO_ENTRIES(btree, r->space_avail) - page->entries;
-	r->recno += entry;
-
-	/* Walk any append list. */
-	for (ins =
-	    WT_SKIP_FIRST(WT_COL_APPEND(page));; ins = WT_SKIP_NEXT(ins)) {
-		if (ins == NULL) {
-			/*
-			 * If the page split, instantiate any missing records in
-			 * the page's name space. (Imagine record 98 is
-			 * transactionally visible, 99 wasn't created or is not
-			 * yet visible, 100 is visible. Then the page splits and
-			 * record 100 moves to another page. When we reconcile
-			 * the original page, we write record 98, then we don't
-			 * see record 99 for whatever reason. If we've moved
-			 * record 100, we don't know to write a deleted record
-			 * 99 on the page.)
-			 *
-			 * The record number recorded during the split is the
-			 * first key on the split page, that is, one larger than
-			 * the last key on this page, we have to decrement it.
-			 */
-			if ((recno =
-			    page->modify->mod_col_split_recno) == WT_RECNO_OOB)
-				break;
-			recno -= 1;
-
-			/*
-			 * The following loop assumes records to write, and the
-			 * previous key might have been visible.
-			 */
-			if (r->recno > recno)
-				break;
-			upd = NULL;
-		} else {
-			WT_RET(__rec_txn_read(
-			    session, r, ins, NULL, NULL, NULL, &upd));
-			recno = WT_INSERT_RECNO(ins);
-		}
-		for (;;) {
-			/*
-			 * The application may have inserted records which left
-			 * gaps in the name space.
-			 */
-			for (;
-			    nrecs > 0 && r->recno < recno;
-			    --nrecs, ++entry, ++r->recno)
-				__bit_setv(
-				    r->first_free, entry, btree->bitcnt, 0);
-
-			if (nrecs > 0) {
-				__bit_setv(r->first_free, entry, btree->bitcnt,
-				    upd == NULL ? 0 : *upd->data);
-				--nrecs;
-				++entry;
-				++r->recno;
-				break;
-			}
-
-			/*
-			 * If everything didn't fit, update the counters and
-			 * split.
-			 *
-			 * Boundary: split or write the page.
-			 *
-			 * No need to have a minimum split size boundary, all
-			 * pages are filled 100% except the last, allowing it to
-			 * grow in the future.
-			 */
-			__rec_incr(session, r, entry,
-			    __bitstr_size((size_t)entry * btree->bitcnt));
-			WT_RET(__rec_split(session, r, 0));
-
-			/* Calculate the number of entries per page. */
-			entry = 0;
-			nrecs = WT_FIX_BYTES_TO_ENTRIES(btree, r->space_avail);
-		}
-
-		/*
-		 * Execute this loop once without an insert item to catch any
-		 * missing records due to a split, then quit.
-		 */
-		if (ins == NULL)
-			break;
-	}
-
-	/* Update the counters. */
-	__rec_incr(
-	    session, r, entry, __bitstr_size((size_t)entry * btree->bitcnt));
-
-	/* Write the remnant page. */
-	return (__rec_split_finish(session, r));
-}
-
-/*
- * __rec_col_fix_slvg --
- *	Reconcile a fixed-width, column-store leaf page created during salvage.
- */
-static int
-__rec_col_fix_slvg(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_REF *pageref, WT_SALVAGE_COOKIE *salvage)
-{
-	WT_BTREE *btree;
-	WT_PAGE *page;
-	uint64_t page_start, page_take;
-	uint32_t entry, nrecs;
-
-	btree = S2BT(session);
-	page = pageref->page;
-
-	/*
-	 * !!!
-	 * It's vanishingly unlikely and probably impossible for fixed-length
-	 * column-store files to have overlapping key ranges.  It's possible
-	 * for an entire key range to go missing (if a page is corrupted and
-	 * lost), but because pages can't split, it shouldn't be possible to
-	 * find pages where the key ranges overlap.  That said, we check for
-	 * it during salvage and clean up after it here because it doesn't
-	 * cost much and future column-store formats or operations might allow
-	 * for fixed-length format ranges to overlap during salvage, and I
-	 * don't want to have to retrofit the code later.
-	 */
-	WT_RET(__rec_split_init(
-	    session, r, page, pageref->ref_recno, btree->maxleafpage));
-
-	/* We may not be taking all of the entries on the original page. */
-	page_take = salvage->take == 0 ? page->entries : salvage->take;
-	page_start = salvage->skip == 0 ? 0 : salvage->skip;
-
-	/* Calculate the number of entries per page. */
-	entry = 0;
-	nrecs = WT_FIX_BYTES_TO_ENTRIES(btree, r->space_avail);
-
-	for (; nrecs > 0 && salvage->missing > 0;
-	    --nrecs, --salvage->missing, ++entry)
-		__bit_setv(r->first_free, entry, btree->bitcnt, 0);
-
-	for (; nrecs > 0 && page_take > 0;
-	    --nrecs, --page_take, ++page_start, ++entry)
-		__bit_setv(r->first_free, entry, btree->bitcnt,
-		    __bit_getv(page->pg_fix_bitf,
-			(uint32_t)page_start, btree->bitcnt));
-
-	r->recno += entry;
-	__rec_incr(session, r, entry,
-	    __bitstr_size((size_t)entry * btree->bitcnt));
-
-	/*
-	 * We can't split during salvage -- if everything didn't fit, it's
-	 * all gone wrong.
-	 */
-	if (salvage->missing != 0 || page_take != 0)
-		WT_PANIC_RET(session, WT_PANIC,
-		    "%s page too large, attempted split during salvage",
-		    __wt_page_type_string(page->type));
-
-	/* Write the page. */
-	return (__rec_split_finish(session, r));
-}
-
-/*
- * __rec_col_var_helper --
- *	Create a column-store variable length record cell and write it onto a
- *	page.
- */
-static int
-__rec_col_var_helper(WT_SESSION_IMPL *session, WT_RECONCILE *r,
-    WT_SALVAGE_COOKIE *salvage,
-    WT_ITEM *value, bool deleted, uint8_t overflow_type, uint64_t rle)
-{
-	WT_BTREE *btree;
-	WT_KV *val;
-
-	btree = S2BT(session);
-
-	val = &r->v;
-
-	/*
-	 * Occasionally, salvage needs to discard records from the beginning or
-	 * end of the page, and because the items may be part of a RLE cell, do
-	 * the adjustments here. It's not a mistake we don't bother telling
-	 * our caller we've handled all the records from the page we care about,
-	 * and can quit processing the page: salvage is a rare operation and I
-	 * don't want to complicate our caller's loop.
-	 */
-	if (salvage != NULL) {
-		if (salvage->done)
-			return (0);
-		if (salvage->skip != 0) {
-			if (rle <= salvage->skip) {
-				salvage->skip -= rle;
-				return (0);
-			}
-			rle -= salvage->skip;
-			salvage->skip = 0;
-		}
-		if (salvage->take != 0) {
-			if (rle <= salvage->take)
-				salvage->take -= rle;
-			else {
-				rle = salvage->take;
-				salvage->take = 0;
-			}
-			if (salvage->take == 0)
-				salvage->done = true;
-		}
-	}
-
-	if (deleted) {
-		val->cell_len = __wt_cell_pack_del(&val->cell, rle);
-		val->buf.data = NULL;
-		val->buf.size = 0;
-		val->len = val->cell_len;
-	} else if (overflow_type) {
-		val->cell_len = __wt_cell_pack_ovfl(
-		    &val->cell, overflow_type, rle, value->size);
-		val->buf.data = value->data;
-		val->buf.size = value->size;
-		val->len = val->cell_len + value->size;
-	} else
-		WT_RET(__rec_cell_build_val(
-		    session, r, value->data, value->size, rle));
-
-	/* Boundary: split or write the page. */
-	if (__rec_need_split(r, val->len))
-		WT_RET(__rec_split_crossing_bnd(session, r, val->len));
-
-	/* Copy the value onto the page. */
-	if (!deleted && !overflow_type && btree->dictionary)
-		WT_RET(__rec_dict_replace(session, r, rle, val));
-	__rec_copy_incr(session, r, val);
-
-	/* Update the starting record number in case we split. */
-	r->recno += rle;
-
-	return (0);
-}
-
-/*
- * __rec_col_var --
- *	Reconcile a variable-width column-store leaf page.
- */
-static int
-__rec_col_var(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_REF *pageref, WT_SALVAGE_COOKIE *salvage)
-{
-	enum { OVFL_IGNORE, OVFL_UNUSED, OVFL_USED } ovfl_state;
-	WT_BTREE *btree;
-	WT_CELL *cell;
-	WT_CELL_UNPACK *vpack, _vpack;
-	WT_COL *cip;
-	WT_CURSOR_BTREE *cbt;
-	WT_DECL_ITEM(orig);
-	WT_DECL_RET;
-	WT_INSERT *ins;
-	WT_ITEM *last;
-	WT_PAGE *page;
-	WT_UPDATE *upd;
-	uint64_t n, nrepeat, repeat_count, rle, skip, src_recno;
-	uint32_t i, size;
-	bool deleted, last_deleted, orig_deleted, update_no_copy;
-	const void *data;
-
-	btree = S2BT(session);
-	page = pageref->page;
-	last = r->last;
-	vpack = &_vpack;
-	cbt = &r->update_modify_cbt;
-
-	WT_RET(__rec_split_init(session,
-	    r, page, pageref->ref_recno, btree->maxleafpage_precomp));
-
-	WT_RET(__wt_scr_alloc(session, 0, &orig));
-	data = NULL;
-	size = 0;
-	upd = NULL;
-
-	/*
-	 * The salvage code may be calling us to reconcile a page where there
-	 * were missing records in the column-store name space.  If taking the
-	 * first record from on the page, it might be a deleted record, so we
-	 * have to give the RLE code a chance to figure that out.  Else, if
-	 * not taking the first record from the page, write a single element
-	 * representing the missing records onto a new page.  (Don't pass the
-	 * salvage cookie to our helper function in this case, we're handling
-	 * one of the salvage cookie fields on our own, and we don't need the
-	 * helper function's assistance.)
-	 */
-	rle = 0;
-	last_deleted = false;
-	if (salvage != NULL && salvage->missing != 0) {
-		if (salvage->skip == 0) {
-			rle = salvage->missing;
-			last_deleted = true;
-
-			/*
-			 * Correct the number of records we're going to "take",
-			 * pretending the missing records were on the page.
-			 */
-			salvage->take += salvage->missing;
-		} else
-			WT_ERR(__rec_col_var_helper(session,
-			    r, NULL, NULL, true, false, salvage->missing));
-	}
-
-	/*
-	 * We track two data items through this loop: the previous (last) item
-	 * and the current item: if the last item is the same as the current
-	 * item, we increment the RLE count for the last item; if the last item
-	 * is different from the current item, we write the last item onto the
-	 * page, and replace it with the current item.  The r->recno counter
-	 * tracks records written to the page, and is incremented by the helper
-	 * function immediately after writing records to the page.  The record
-	 * number of our source record, that is, the current item, is maintained
-	 * in src_recno.
-	 */
-	src_recno = r->recno + rle;
-
-	/* For each entry in the in-memory page... */
-	WT_COL_FOREACH(page, cip, i) {
-		ovfl_state = OVFL_IGNORE;
-		if ((cell = WT_COL_PTR(page, cip)) == NULL) {
-			nrepeat = 1;
-			ins = NULL;
-			orig_deleted = true;
-		} else {
-			__wt_cell_unpack(cell, vpack);
-			nrepeat = __wt_cell_rle(vpack);
-			ins = WT_SKIP_FIRST(WT_COL_UPDATE(page, cip));
-
-			/*
-			 * If the original value is "deleted", there's no value
-			 * to compare, we're done.
-			 */
-			orig_deleted = vpack->type == WT_CELL_DEL;
-			if (orig_deleted)
-				goto record_loop;
-
-			/*
-			 * Overflow items are tricky: we don't know until we're
-			 * finished processing the set of values if we need the
-			 * overflow value or not.  If we don't use the overflow
-			 * item at all, we have to discard it from the backing
-			 * file, otherwise we'll leak blocks on the checkpoint.
-			 * That's safe because if the backing overflow value is
-			 * still needed by any running transaction, we'll cache
-			 * a copy in the update list.
-			 *
-			 * Regardless, we avoid copying in overflow records: if
-			 * there's a WT_INSERT entry that modifies a reference
-			 * counted overflow record, we may have to write copies
-			 * of the overflow record, and in that case we'll do the
-			 * comparisons, but we don't read overflow items just to
-			 * see if they match records on either side.
-			 */
-			if (vpack->ovfl) {
-				ovfl_state = OVFL_UNUSED;
-				goto record_loop;
-			}
-
-			/*
-			 * If data is Huffman encoded, we have to decode it in
-			 * order to compare it with the last item we saw, which
-			 * may have been an update string.  This guarantees we
-			 * find every single pair of objects we can RLE encode,
-			 * including applications updating an existing record
-			 * where the new value happens (?) to match a Huffman-
-			 * encoded value in a previous or next record.
-			 */
-			WT_ERR(__wt_dsk_cell_data_ref(
-			    session, WT_PAGE_COL_VAR, vpack, orig));
-		}
-
-record_loop:	/*
-		 * Generate on-page entries: loop repeat records, looking for
-		 * WT_INSERT entries matching the record number.  The WT_INSERT
-		 * lists are in sorted order, so only need check the next one.
-		 */
-		for (n = 0;
-		    n < nrepeat; n += repeat_count, src_recno += repeat_count) {
-			upd = NULL;
-			if (ins != NULL && WT_INSERT_RECNO(ins) == src_recno) {
-				WT_ERR(__rec_txn_read(
-				    session, r, ins, cip, vpack, NULL, &upd));
-				ins = WT_SKIP_NEXT(ins);
-			}
-
-			update_no_copy = true;	/* No data copy */
-			repeat_count = 1;	/* Single record */
-			deleted = false;
-
-			if (upd != NULL) {
-				switch (upd->type) {
-				case WT_UPDATE_MODIFY:
-					cbt->slot = WT_COL_SLOT(page, cip);
-					WT_ERR(__wt_value_return_upd(
-					    session, cbt, upd,
-					    F_ISSET(r, WT_REC_VISIBLE_ALL)));
-					data = cbt->iface.value.data;
-					size = (uint32_t)cbt->iface.value.size;
-					update_no_copy = false;
-					break;
-				case WT_UPDATE_STANDARD:
-					data = upd->data;
-					size = upd->size;
-					break;
-				case WT_UPDATE_TOMBSTONE:
-					deleted = true;
-					break;
-				WT_ILLEGAL_VALUE_ERR(session, upd->type);
-				}
-			} else if (vpack->raw == WT_CELL_VALUE_OVFL_RM) {
-				/*
-				 * If doing an update save and restore, and the
-				 * underlying value is a removed overflow value,
-				 * we end up here.
-				 *
-				 * If necessary, when the overflow value was
-				 * originally removed, reconciliation appended
-				 * a globally visible copy of the value to the
-				 * key's update list, meaning the on-page item
-				 * isn't accessed after page re-instantiation.
-				 *
-				 * Assert the case.
-				 */
-				WT_ASSERT(session,
-				    F_ISSET(r, WT_REC_UPDATE_RESTORE));
-
-				/*
-				 * The on-page value will never be accessed,
-				 * write a placeholder record.
-				 */
-				data = "ovfl-unused";
-				size = WT_STORE_SIZE(strlen("ovfl-unused"));
-			} else {
-				update_no_copy = false;	/* Maybe data copy */
-
-				/*
-				 * The repeat count is the number of records up
-				 * to the next WT_INSERT record, or up to the
-				 * end of the entry if we have no more WT_INSERT
-				 * records.
-				 */
-				if (ins == NULL)
-					repeat_count = nrepeat - n;
-				else
-					repeat_count =
-					    WT_INSERT_RECNO(ins) - src_recno;
-
-				deleted = orig_deleted;
-				if (deleted)
-					goto compare;
-
-				/*
-				 * If we are handling overflow items, use the
-				 * overflow item itself exactly once, after
-				 * which we have to copy it into a buffer and
-				 * from then on use a complete copy because we
-				 * are re-creating a new overflow record each
-				 * time.
-				 */
-				switch (ovfl_state) {
-				case OVFL_UNUSED:
-					/*
-					 * An as-yet-unused overflow item.
-					 *
-					 * We're going to copy the on-page cell,
-					 * write out any record we're tracking.
-					 */
-					if (rle != 0) {
-						WT_ERR(__rec_col_var_helper(
-						    session, r, salvage, last,
-						    last_deleted, 0, rle));
-						rle = 0;
-					}
-
-					last->data = vpack->data;
-					last->size = vpack->size;
-					WT_ERR(__rec_col_var_helper(
-					    session, r, salvage, last, false,
-					    WT_CELL_VALUE_OVFL, repeat_count));
-
-					/* Track if page has overflow items. */
-					r->ovfl_items = true;
-
-					ovfl_state = OVFL_USED;
-					continue;
-				case OVFL_USED:
-					/*
-					 * Original is an overflow item; we used
-					 * it for a key and now we need another
-					 * copy; read it into memory.
-					 */
-					WT_ERR(__wt_dsk_cell_data_ref(session,
-					    WT_PAGE_COL_VAR, vpack, orig));
-
-					ovfl_state = OVFL_IGNORE;
-					/* FALLTHROUGH */
-				case OVFL_IGNORE:
-					/*
-					 * Original is an overflow item and we
-					 * were forced to copy it into memory,
-					 * or the original wasn't an overflow
-					 * item; use the data copied into orig.
-					 */
-					data = orig->data;
-					size = (uint32_t)orig->size;
-					break;
-				}
-			}
-
-compare:		/*
-			 * If we have a record against which to compare, and
-			 * the records compare equal, increment the rle counter
-			 * and continue.  If the records don't compare equal,
-			 * output the last record and swap the last and current
-			 * buffers: do NOT update the starting record number,
-			 * we've been doing that all along.
-			 */
-			if (rle != 0) {
-				if ((deleted && last_deleted) ||
-				    (!last_deleted && !deleted &&
-				    last->size == size &&
-				    memcmp(last->data, data, size) == 0)) {
-					rle += repeat_count;
-					continue;
-				}
-				WT_ERR(__rec_col_var_helper(session, r,
-				    salvage, last, last_deleted, 0, rle));
-			}
-
-			/*
-			 * Swap the current/last state.
-			 *
-			 * Reset RLE counter and turn on comparisons.
-			 */
-			if (!deleted) {
-				/*
-				 * We can't simply assign the data values into
-				 * the last buffer because they may have come
-				 * from a copy built from an encoded/overflow
-				 * cell and creating the next record is going
-				 * to overwrite that memory.  Check, because
-				 * encoded/overflow cells aren't that common
-				 * and we'd like to avoid the copy.  If data
-				 * was taken from the current unpack structure
-				 * (which points into the page), or was taken
-				 * from an update structure, we can just use
-				 * the pointers, they're not moving.
-				 */
-				if (data == vpack->data || update_no_copy) {
-					last->data = data;
-					last->size = size;
-				} else
-					WT_ERR(__wt_buf_set(
-					    session, last, data, size));
-			}
-			last_deleted = deleted;
-			rle = repeat_count;
-		}
-
-		/*
-		 * The first time we find an overflow record we never used,
-		 * discard the underlying blocks, they're no longer useful.
-		 */
-		if (ovfl_state == OVFL_UNUSED &&
-		    vpack->raw != WT_CELL_VALUE_OVFL_RM)
-			WT_ERR(__wt_ovfl_remove(
-			    session, page, vpack, F_ISSET(r, WT_REC_EVICT)));
-	}
-
-	/* Walk any append list. */
-	for (ins =
-	    WT_SKIP_FIRST(WT_COL_APPEND(page));; ins = WT_SKIP_NEXT(ins)) {
-		if (ins == NULL) {
-			/*
-			 * If the page split, instantiate any missing records in
-			 * the page's name space. (Imagine record 98 is
-			 * transactionally visible, 99 wasn't created or is not
-			 * yet visible, 100 is visible. Then the page splits and
-			 * record 100 moves to another page. When we reconcile
-			 * the original page, we write record 98, then we don't
-			 * see record 99 for whatever reason. If we've moved
-			 * record 100, we don't know to write a deleted record
-			 * 99 on the page.)
-			 *
-			 * Assert the recorded record number is past the end of
-			 * the page.
-			 *
-			 * The record number recorded during the split is the
-			 * first key on the split page, that is, one larger than
-			 * the last key on this page, we have to decrement it.
-			 */
-			if ((n = page->
-			    modify->mod_col_split_recno) == WT_RECNO_OOB)
-				break;
-			WT_ASSERT(session, n >= src_recno);
-			n -= 1;
-
-			upd = NULL;
-		} else {
-			WT_ERR(__rec_txn_read(
-			    session, r, ins, NULL, NULL, NULL, &upd));
-			n = WT_INSERT_RECNO(ins);
-		}
-		while (src_recno <= n) {
-			deleted = false;
-			update_no_copy = true;
-
-			/*
-			 * The application may have inserted records which left
-			 * gaps in the name space, and these gaps can be huge.
-			 * If we're in a set of deleted records, skip the boring
-			 * part.
-			 */
-			if (src_recno < n) {
-				deleted = true;
-				if (last_deleted) {
-					/*
-					 * The record adjustment is decremented
-					 * by one so we can naturally fall into
-					 * the RLE accounting below, where we
-					 * increment rle by one, then continue
-					 * in the outer loop, where we increment
-					 * src_recno by one.
-					 */
-					skip = (n - src_recno) - 1;
-					rle += skip;
-					src_recno += skip;
-				}
-			} else if (upd == NULL)
-				deleted = true;
-			else
-				switch (upd->type) {
-				case WT_UPDATE_MODIFY:
-					/*
-					 * Impossible slot, there's no backing
-					 * on-page item.
-					 */
-					cbt->slot = UINT32_MAX;
-					WT_ERR(__wt_value_return_upd(
-					    session, cbt, upd,
-					    F_ISSET(r, WT_REC_VISIBLE_ALL)));
-					data = cbt->iface.value.data;
-					size = (uint32_t)cbt->iface.value.size;
-					update_no_copy = false;
-					break;
-				case WT_UPDATE_STANDARD:
-					data = upd->data;
-					size = upd->size;
-					break;
-				case WT_UPDATE_TOMBSTONE:
-					deleted = true;
-					break;
-				WT_ILLEGAL_VALUE_ERR(session, upd->type);
-				}
-
-			/*
-			 * Handle RLE accounting and comparisons -- see comment
-			 * above, this code fragment does the same thing.
-			 */
-			if (rle != 0) {
-				if ((deleted && last_deleted) ||
-				    (!last_deleted && !deleted &&
-				    last->size == size &&
-				    memcmp(last->data, data, size) == 0)) {
-					++rle;
-					goto next;
-				}
-				WT_ERR(__rec_col_var_helper(session, r,
-				    salvage, last, last_deleted, 0, rle));
-			}
-
-			/*
-			 * Swap the current/last state. We can't simply assign
-			 * the data values into the last buffer because they may
-			 * be a temporary copy built from a chain of modified
-			 * updates and creating the next record will overwrite
-			 * that memory. Check, we'd like to avoid the copy. If
-			 * data was taken from an update structure, we can just
-			 * use the pointers, they're not moving.
-			 */
-			if (!deleted) {
-				if (update_no_copy) {
-					last->data = data;
-					last->size = size;
-				} else
-					WT_ERR(__wt_buf_set(
-					    session, last, data, size));
-			}
-
-			/* Ready for the next loop, reset the RLE counter. */
-			last_deleted = deleted;
-			rle = 1;
-
-			/*
-			 * Move to the next record. It's not a simple increment
-			 * because if it's the maximum record, incrementing it
-			 * wraps to 0 and this turns into an infinite loop.
-			 */
-next:			if (src_recno == UINT64_MAX)
-				break;
-			++src_recno;
-		}
-
-		/*
-		 * Execute this loop once without an insert item to catch any
-		 * missing records due to a split, then quit.
-		 */
-		if (ins == NULL)
-			break;
-	}
-
-	/* If we were tracking a record, write it. */
-	if (rle != 0)
-		WT_ERR(__rec_col_var_helper(
-		    session, r, salvage, last, last_deleted, 0, rle));
-
-	/* Write the remnant page. */
-	ret = __rec_split_finish(session, r);
-
-err:	__wt_scr_free(session, &orig);
-	return (ret);
-}
-
-/*
- * __rec_row_int --
- *	Reconcile a row-store internal page.
- */
-static int
-__rec_row_int(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_PAGE *page)
-{
-	WT_ADDR *addr;
-	WT_BTREE *btree;
-	WT_CELL *cell;
-	WT_CELL_UNPACK *kpack, _kpack, *vpack, _vpack;
-	WT_CHILD_STATE state;
-	WT_DECL_RET;
-	WT_IKEY *ikey;
-	WT_KV *key, *val;
-	WT_PAGE *child;
-	WT_REF *ref;
-	size_t size;
-	u_int vtype;
-	bool hazard, key_onpage_ovfl, ovfl_key;
-	const void *p;
-
-	btree = S2BT(session);
-	child = NULL;
-	hazard = false;
-
-	key = &r->k;
-	kpack = &_kpack;
-	WT_CLEAR(*kpack);	/* -Wuninitialized */
-	val = &r->v;
-	vpack = &_vpack;
-	WT_CLEAR(*vpack);	/* -Wuninitialized */
-
-	ikey = NULL;		/* -Wuninitialized */
-	cell = NULL;
-	key_onpage_ovfl = false;
-
-	WT_RET(__rec_split_init(
-	    session, r, page, 0, btree->maxintlpage_precomp));
-
-	/*
-	 * Ideally, we'd never store the 0th key on row-store internal pages
-	 * because it's never used during tree search and there's no reason
-	 * to waste the space.  The problem is how we do splits: when we split,
-	 * we've potentially picked out several "split points" in the buffer
-	 * which is overflowing the maximum page size, and when the overflow
-	 * happens, we go back and physically split the buffer, at those split
-	 * points, into new pages.  It would be both difficult and expensive
-	 * to re-process the 0th key at each split point to be an empty key,
-	 * so we don't do that.  However, we are reconciling an internal page
-	 * for whatever reason, and the 0th key is known to be useless.  We
-	 * truncate the key to a single byte, instead of removing it entirely,
-	 * it simplifies various things in other parts of the code (we don't
-	 * have to special case transforming the page from its disk image to
-	 * its in-memory version, for example).
-	 */
-	r->cell_zero = true;
-
-	/* For each entry in the in-memory page... */
-	WT_INTL_FOREACH_BEGIN(session, page, ref) {
-		/*
-		 * There are different paths if the key is an overflow item vs.
-		 * a straight-forward on-page value. If an overflow item, we
-		 * would have instantiated it, and we can use that fact to set
-		 * things up.
-		 *
-		 * Note the cell reference and unpacked key cell are available
-		 * only in the case of an instantiated, off-page key, we don't
-		 * bother setting them if that's not possible.
-		 */
-		if (F_ISSET_ATOMIC(page, WT_PAGE_OVERFLOW_KEYS)) {
-			cell = NULL;
-			key_onpage_ovfl = false;
-			ikey = __wt_ref_key_instantiated(ref);
-			if (ikey != NULL && ikey->cell_offset != 0) {
-				cell =
-				    WT_PAGE_REF_OFFSET(page, ikey->cell_offset);
-				__wt_cell_unpack(cell, kpack);
-				key_onpage_ovfl = kpack->ovfl &&
-				    kpack->raw != WT_CELL_KEY_OVFL_RM;
-			}
-		}
-
-		WT_ERR(__rec_child_modify(session, r, ref, &hazard, &state));
-		addr = ref->addr;
-		child = ref->page;
-
-		switch (state) {
-		case WT_CHILD_IGNORE:
-			/*
-			 * Ignored child.
-			 *
-			 * Overflow keys referencing pages we're not writing are
-			 * no longer useful, schedule them for discard.  Don't
-			 * worry about instantiation, internal page keys are
-			 * always instantiated.  Don't worry about reuse,
-			 * reusing this key in this reconciliation is unlikely.
-			 */
-			if (key_onpage_ovfl)
-				WT_ERR(__wt_ovfl_discard_add(
-				    session, page, kpack->cell));
-			WT_CHILD_RELEASE_ERR(session, hazard, ref);
-			continue;
-
-		case WT_CHILD_MODIFIED:
-			/*
-			 * Modified child.  Empty pages are merged into the
-			 * parent and discarded.
-			 */
-			switch (child->modify->rec_result) {
-			case WT_PM_REC_EMPTY:
-				/*
-				 * Overflow keys referencing empty pages are no
-				 * longer useful, schedule them for discard.
-				 * Don't worry about instantiation, internal
-				 * page keys are always instantiated.  Don't
-				 * worry about reuse, reusing this key in this
-				 * reconciliation is unlikely.
-				 */
-				if (key_onpage_ovfl)
-					WT_ERR(__wt_ovfl_discard_add(
-					    session, page, kpack->cell));
-				WT_CHILD_RELEASE_ERR(session, hazard, ref);
-				continue;
-			case WT_PM_REC_MULTIBLOCK:
-				/*
-				 * Overflow keys referencing split pages are no
-				 * longer useful (the split page's key is the
-				 * interesting key); schedule them for discard.
-				 * Don't worry about instantiation, internal
-				 * page keys are always instantiated.  Don't
-				 * worry about reuse, reusing this key in this
-				 * reconciliation is unlikely.
-				 */
-				if (key_onpage_ovfl)
-					WT_ERR(__wt_ovfl_discard_add(
-					    session, page, kpack->cell));
-
-				WT_ERR(__rec_row_merge(session, r, child));
-				WT_CHILD_RELEASE_ERR(session, hazard, ref);
-				continue;
-			case WT_PM_REC_REPLACE:
-				/*
-				 * If the page is replaced, the page's modify
-				 * structure has the page's address.
-				 */
-				addr = &child->modify->mod_replace;
-				break;
-			WT_ILLEGAL_VALUE_ERR(
-			    session, child->modify->rec_result);
-			}
-			break;
-		case WT_CHILD_ORIGINAL:
-			/* Original child. */
-			break;
-		case WT_CHILD_PROXY:
-			/* Deleted child where we write a proxy cell. */
-			break;
-		}
-
-		/*
-		 * Build the value cell, the child page's address.  Addr points
-		 * to an on-page cell or an off-page WT_ADDR structure. There's
-		 * a special cell type in the case of page deletion requiring
-		 * a proxy cell, otherwise use the information from the addr or
-		 * original cell.
-		 */
-		if (__wt_off_page(page, addr)) {
-			p = addr->addr;
-			size = addr->size;
-			vtype = state == WT_CHILD_PROXY ?
-			    WT_CELL_ADDR_DEL : __rec_vtype(addr);
-		} else {
-			__wt_cell_unpack(ref->addr, vpack);
-			p = vpack->data;
-			size = vpack->size;
-			vtype = state == WT_CHILD_PROXY ?
-			    WT_CELL_ADDR_DEL : (u_int)vpack->raw;
-		}
-		__rec_cell_build_addr(session, r, p, size, vtype, WT_RECNO_OOB);
-		WT_CHILD_RELEASE_ERR(session, hazard, ref);
-
-		/*
-		 * Build key cell.
-		 * Truncate any 0th key, internal pages don't need 0th keys.
-		 */
-		if (key_onpage_ovfl) {
-			key->buf.data = cell;
-			key->buf.size = __wt_cell_total_len(kpack);
-			key->cell_len = 0;
-			key->len = key->buf.size;
-			ovfl_key = true;
-		} else {
-			__wt_ref_key(page, ref, &p, &size);
-			WT_ERR(__rec_cell_build_int_key(
-			    session, r, p, r->cell_zero ? 1 : size, &ovfl_key));
-		}
-		r->cell_zero = false;
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, key->len + val->len)) {
-			/*
-			 * In one path above, we copied address blocks from the
-			 * page rather than building the actual key. In that
-			 * case, we have to build the key now because we are
-			 * about to promote it.
-			 */
-			if (key_onpage_ovfl) {
-				WT_ERR(__wt_buf_set(session, r->cur,
-				    WT_IKEY_DATA(ikey), ikey->size));
-				key_onpage_ovfl = false;
-			}
-
-			WT_ERR(__rec_split_crossing_bnd(
-			    session, r, key->len + val->len));
-		}
-
-		/* Copy the key and value onto the page. */
-		__rec_copy_incr(session, r, key);
-		__rec_copy_incr(session, r, val);
-
-		/* Update compression state. */
-		__rec_key_state_update(r, ovfl_key);
-	} WT_INTL_FOREACH_END;
-
-	/* Write the remnant page. */
-	return (__rec_split_finish(session, r));
-
-err:	WT_CHILD_RELEASE(session, hazard, ref);
-	return (ret);
-}
-
-/*
- * __rec_row_merge --
- *	Merge in a split page.
- */
-static int
-__rec_row_merge(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_PAGE *page)
-{
-	WT_ADDR *addr;
-	WT_KV *key, *val;
-	WT_MULTI *multi;
-	WT_PAGE_MODIFY *mod;
-	uint32_t i;
-	bool ovfl_key;
-
-	mod = page->modify;
-
-	key = &r->k;
-	val = &r->v;
-
-	/* For each entry in the split array... */
-	for (multi = mod->mod_multi,
-	    i = 0; i < mod->mod_multi_entries; ++multi, ++i) {
-		/* Build the key and value cells. */
-		WT_RET(__rec_cell_build_int_key(session, r,
-		    WT_IKEY_DATA(multi->key.ikey),
-		    r->cell_zero ? 1 : multi->key.ikey->size, &ovfl_key));
-		r->cell_zero = false;
-
-		addr = &multi->addr;
-		__rec_cell_build_addr(session, r,
-		    addr->addr, addr->size, __rec_vtype(addr), WT_RECNO_OOB);
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, key->len + val->len))
-			WT_RET(__rec_split_crossing_bnd(
-			    session, r, key->len + val->len));
-
-		/* Copy the key and value onto the page. */
-		__rec_copy_incr(session, r, key);
-		__rec_copy_incr(session, r, val);
-
-		/* Update compression state. */
-		__rec_key_state_update(r, ovfl_key);
-	}
-	return (0);
-}
-
-/*
- * __rec_row_leaf --
- *	Reconcile a row-store leaf page.
- */
-static int
-__rec_row_leaf(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_PAGE *page, WT_SALVAGE_COOKIE *salvage)
-{
-	WT_BTREE *btree;
-	WT_CELL *cell;
-	WT_CELL_UNPACK *kpack, _kpack, *vpack, _vpack;
-	WT_CURSOR_BTREE *cbt;
-	WT_DECL_ITEM(tmpkey);
-	WT_DECL_ITEM(tmpval);
-	WT_DECL_RET;
-	WT_IKEY *ikey;
-	WT_INSERT *ins;
-	WT_KV *key, *val;
-	WT_ROW *rip;
-	WT_UPDATE *upd;
-	size_t size;
-	uint64_t slvg_skip;
-	uint32_t i;
-	bool dictionary, key_onpage_ovfl, ovfl_key;
-	void *copy;
-	const void *p;
-
-	btree = S2BT(session);
-	cbt = &r->update_modify_cbt;
-	slvg_skip = salvage == NULL ? 0 : salvage->skip;
-
-	key = &r->k;
-	val = &r->v;
-	vpack = &_vpack;
-
-	WT_RET(__rec_split_init(
-	    session, r, page, 0, btree->maxleafpage_precomp));
-
-	/*
-	 * Write any K/V pairs inserted into the page before the first from-disk
-	 * key on the page.
-	 */
-	if ((ins = WT_SKIP_FIRST(WT_ROW_INSERT_SMALLEST(page))) != NULL)
-		WT_RET(__rec_row_leaf_insert(session, r, ins));
-
-	/*
-	 * Temporary buffers in which to instantiate any uninstantiated keys
-	 * or value items we need.
-	 */
-	WT_ERR(__wt_scr_alloc(session, 0, &tmpkey));
-	WT_ERR(__wt_scr_alloc(session, 0, &tmpval));
-
-	/* For each entry in the page... */
-	WT_ROW_FOREACH(page, rip, i) {
-		/*
-		 * The salvage code, on some rare occasions, wants to reconcile
-		 * a page but skip some leading records on the page.  Because
-		 * the row-store leaf reconciliation function copies keys from
-		 * the original disk page, this is non-trivial -- just changing
-		 * the in-memory pointers isn't sufficient, we have to change
-		 * the WT_CELL structures on the disk page, too.  It's ugly, but
-		 * we pass in a value that tells us how many records to skip in
-		 * this case.
-		 */
-		if (slvg_skip != 0) {
-			--slvg_skip;
-			continue;
-		}
-
-		/*
-		 * Figure out the key: set any cell reference (and unpack it),
-		 * set any instantiated key reference.
-		 */
-		copy = WT_ROW_KEY_COPY(rip);
-		(void)__wt_row_leaf_key_info(
-		    page, copy, &ikey, &cell, NULL, NULL);
-		if (cell == NULL)
-			kpack = NULL;
-		else {
-			kpack = &_kpack;
-			__wt_cell_unpack(cell, kpack);
-		}
-
-		/* Unpack the on-page value cell, and look for an update. */
-		__wt_row_leaf_value_cell(page, rip, NULL, vpack);
-		WT_ERR(__rec_txn_read(
-		    session, r, NULL, rip, vpack, NULL, &upd));
-
-		/* Build value cell. */
-		dictionary = false;
-		if (upd == NULL) {
-			/*
-			 * When the page was read into memory, there may not
-			 * have been a value item.
-			 *
-			 * If there was a value item, check if it's a dictionary
-			 * cell (a copy of another item on the page).  If it's a
-			 * copy, we have to create a new value item as the old
-			 * item might have been discarded from the page.
-			 */
-			if (vpack->raw == WT_CELL_VALUE_COPY) {
-				/* If the item is Huffman encoded, decode it. */
-				if (btree->huffman_value == NULL) {
-					p = vpack->data;
-					size = vpack->size;
-				} else {
-					WT_ERR(__wt_huffman_decode(session,
-					    btree->huffman_value,
-					    vpack->data, vpack->size,
-					    tmpval));
-					p = tmpval->data;
-					size = tmpval->size;
-				}
-				WT_ERR(__rec_cell_build_val(
-				    session, r, p, size, (uint64_t)0));
-				dictionary = true;
-			} else if (vpack->raw == WT_CELL_VALUE_OVFL_RM) {
-				/*
-				 * If doing an update save and restore, and the
-				 * underlying value is a removed overflow value,
-				 * we end up here.
-				 *
-				 * If necessary, when the overflow value was
-				 * originally removed, reconciliation appended
-				 * a globally visible copy of the value to the
-				 * key's update list, meaning the on-page item
-				 * isn't accessed after page re-instantiation.
-				 *
-				 * Assert the case.
-				 */
-				WT_ASSERT(session,
-				    F_ISSET(r, WT_REC_UPDATE_RESTORE));
-
-				/*
-				 * If the key is also a removed overflow item,
-				 * don't write anything at all.
-				 *
-				 * We don't have to write anything because the
-				 * code re-instantiating the page gets the key
-				 * to match the saved list of updates from the
-				 * original page.  By not putting the key on
-				 * the page, we'll move the key/value set from
-				 * a row-store leaf page slot to an insert list,
-				 * but that shouldn't matter.
-				 *
-				 * The reason we bother with the test is because
-				 * overflows are expensive to write.  It's hard
-				 * to imagine a real workload where this test is
-				 * worth the effort, but it's a simple test.
-				 */
-				if (kpack != NULL &&
-				    kpack->raw == WT_CELL_KEY_OVFL_RM)
-					goto leaf_insert;
-
-				/*
-				 * The on-page value will never be accessed,
-				 * write a placeholder record.
-				 */
-				WT_ERR(__rec_cell_build_val(session, r,
-				    "ovfl-unused", strlen("ovfl-unused"),
-				    (uint64_t)0));
-			} else {
-				val->buf.data = vpack->cell;
-				val->buf.size = __wt_cell_total_len(vpack);
-				val->cell_len = 0;
-				val->len = val->buf.size;
-
-				/* Track if page has overflow items. */
-				if (vpack->ovfl)
-					r->ovfl_items = true;
-			}
-		} else {
-			/*
-			 * The first time we find an overflow record we're not
-			 * going to use, discard the underlying blocks.
-			 */
-			if (vpack->ovfl && vpack->raw != WT_CELL_VALUE_OVFL_RM)
-				WT_ERR(__wt_ovfl_remove(session,
-				    page, vpack, F_ISSET(r, WT_REC_EVICT)));
-
-			switch (upd->type) {
-			case WT_UPDATE_MODIFY:
-				cbt->slot = WT_ROW_SLOT(page, rip);
-				WT_ERR(__wt_value_return_upd(session, cbt, upd,
-				    F_ISSET(r, WT_REC_VISIBLE_ALL)));
-				WT_ERR(__rec_cell_build_val(session, r,
-				    cbt->iface.value.data,
-				    cbt->iface.value.size, (uint64_t)0));
-				dictionary = true;
-				break;
-			case WT_UPDATE_STANDARD:
-				/*
-				 * If no value, nothing needs to be copied.
-				 * Otherwise, build the value's chunk from the
-				 * update value.
-				 */
-				if (upd->size == 0) {
-					val->buf.data = NULL;
-					val->cell_len =
-					    val->len = val->buf.size = 0;
-				} else {
-					WT_ERR(__rec_cell_build_val(session, r,
-					    upd->data, upd->size,
-					    (uint64_t)0));
-					dictionary = true;
-				}
-				break;
-			case WT_UPDATE_TOMBSTONE:
-				/*
-				 * If this key/value pair was deleted, we're
-				 * done.
-				 *
-				 * Overflow keys referencing discarded values
-				 * are no longer useful, discard the backing
-				 * blocks.  Don't worry about reuse, reusing
-				 * keys from a row-store page reconciliation
-				 * seems unlikely enough to ignore.
-				 */
-				if (kpack != NULL && kpack->ovfl &&
-				    kpack->raw != WT_CELL_KEY_OVFL_RM) {
-					/*
-					 * Keys are part of the name-space, we
-					 * can't remove them from the in-memory
-					 * tree; if an overflow key was deleted
-					 * without being instantiated (for
-					 * example, cursor-based truncation), do
-					 * it now.
-					 */
-					if (ikey == NULL)
-						WT_ERR(__wt_row_leaf_key(
-						    session,
-						    page, rip, tmpkey, true));
-
-					WT_ERR(__wt_ovfl_discard_add(
-					    session, page, kpack->cell));
-				}
-
-				/*
-				 * We aren't actually creating the key so we
-				 * can't use bytes from this key to provide
-				 * prefix information for a subsequent key.
-				 */
-				tmpkey->size = 0;
-
-				/* Proceed with appended key/value pairs. */
-				goto leaf_insert;
-			WT_ILLEGAL_VALUE_ERR(session, upd->type);
-			}
-		}
-
-		/*
-		 * Build key cell.
-		 *
-		 * If the key is an overflow key that hasn't been removed, use
-		 * the original backing blocks.
-		 */
-		key_onpage_ovfl = kpack != NULL &&
-		    kpack->ovfl && kpack->raw != WT_CELL_KEY_OVFL_RM;
-		if (key_onpage_ovfl) {
-			key->buf.data = cell;
-			key->buf.size = __wt_cell_total_len(kpack);
-			key->cell_len = 0;
-			key->len = key->buf.size;
-			ovfl_key = true;
-
-			/*
-			 * We aren't creating a key so we can't use this key as
-			 * a prefix for a subsequent key.
-			 */
-			tmpkey->size = 0;
-
-			/* Track if page has overflow items. */
-			r->ovfl_items = true;
-		} else {
-			/*
-			 * Get the key from the page or an instantiated key, or
-			 * inline building the key from a previous key (it's a
-			 * fast path for simple, prefix-compressed keys), or by
-			 * by building the key from scratch.
-			 */
-			if (__wt_row_leaf_key_info(page, copy,
-			    NULL, &cell, &tmpkey->data, &tmpkey->size))
-				goto build;
-
-			kpack = &_kpack;
-			__wt_cell_unpack(cell, kpack);
-			if (btree->huffman_key == NULL &&
-			    kpack->type == WT_CELL_KEY &&
-			    tmpkey->size >= kpack->prefix) {
-				/*
-				 * The previous clause checked for a prefix of
-				 * zero, which means the temporary buffer must
-				 * have a non-zero size, and it references a
-				 * valid key.
-				 */
-				WT_ASSERT(session, tmpkey->size != 0);
-
-				/*
-				 * Grow the buffer as necessary, ensuring data
-				 * data has been copied into local buffer space,
-				 * then append the suffix to the prefix already
-				 * in the buffer.
-				 *
-				 * Don't grow the buffer unnecessarily or copy
-				 * data we don't need, truncate the item's data
-				 * length to the prefix bytes.
-				 */
-				tmpkey->size = kpack->prefix;
-				WT_ERR(__wt_buf_grow(session,
-				    tmpkey, tmpkey->size + kpack->size));
-				memcpy((uint8_t *)tmpkey->mem + tmpkey->size,
-				    kpack->data, kpack->size);
-				tmpkey->size += kpack->size;
-			} else
-				WT_ERR(__wt_row_leaf_key_copy(
-				    session, page, rip, tmpkey));
-build:
-			WT_ERR(__rec_cell_build_leaf_key(session, r,
-			    tmpkey->data, tmpkey->size, &ovfl_key));
-		}
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, key->len + val->len)) {
-			/*
-			 * If we copied address blocks from the page rather than
-			 * building the actual key, we have to build the key now
-			 * because we are about to promote it.
-			 */
-			if (key_onpage_ovfl) {
-				WT_ERR(__wt_dsk_cell_data_ref(session,
-				    WT_PAGE_ROW_LEAF, kpack, r->cur));
-				WT_NOT_READ(key_onpage_ovfl, false);
-			}
-
-			/*
-			 * Turn off prefix compression until a full key written
-			 * to the new page, and (unless already working with an
-			 * overflow key), rebuild the key without compression.
-			 */
-			if (r->key_pfx_compress_conf) {
-				r->key_pfx_compress = false;
-				if (!ovfl_key)
-					WT_ERR(__rec_cell_build_leaf_key(
-					    session, r, NULL, 0, &ovfl_key));
-			}
-
-			WT_ERR(__rec_split_crossing_bnd(
-			    session, r, key->len + val->len));
-		}
-
-		/* Copy the key/value pair onto the page. */
-		__rec_copy_incr(session, r, key);
-		if (val->len == 0)
-			r->any_empty_value = true;
-		else {
-			r->all_empty_value = false;
-			if (dictionary && btree->dictionary)
-				WT_ERR(__rec_dict_replace(session, r, 0, val));
-			__rec_copy_incr(session, r, val);
-		}
-
-		/* Update compression state. */
-		__rec_key_state_update(r, ovfl_key);
-
-leaf_insert:	/* Write any K/V pairs inserted into the page after this key. */
-		if ((ins = WT_SKIP_FIRST(WT_ROW_INSERT(page, rip))) != NULL)
-			WT_ERR(__rec_row_leaf_insert(session, r, ins));
-	}
-
-	/* Write the remnant page. */
-	ret = __rec_split_finish(session, r);
-
-err:	__wt_scr_free(session, &tmpkey);
-	__wt_scr_free(session, &tmpval);
-	return (ret);
-}
-
-/*
- * __rec_row_leaf_insert --
- *	Walk an insert chain, writing K/V pairs.
- */
-static int
-__rec_row_leaf_insert(WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_INSERT *ins)
-{
-	WT_BTREE *btree;
-	WT_CURSOR_BTREE *cbt;
-	WT_KV *key, *val;
-	WT_UPDATE *upd;
-	bool ovfl_key, upd_saved;
-
-	btree = S2BT(session);
-	cbt = &r->update_modify_cbt;
-
-	key = &r->k;
-	val = &r->v;
-
-	for (; ins != NULL; ins = WT_SKIP_NEXT(ins)) {
-		WT_RET(__rec_txn_read(
-		    session, r, ins, NULL, NULL, &upd_saved, &upd));
-
-		if (upd == NULL) {
-			/*
-			 * If no update is visible but some were saved, check
-			 * for splits.
-			 */
-			if (!upd_saved)
-				continue;
-			if (!__rec_need_split(r, WT_INSERT_KEY_SIZE(ins)))
-				continue;
-
-			/* Copy the current key into place and then split. */
-			WT_RET(__wt_buf_set(session, r->cur,
-			    WT_INSERT_KEY(ins), WT_INSERT_KEY_SIZE(ins)));
-			WT_RET(__rec_split_crossing_bnd(
-			    session, r, WT_INSERT_KEY_SIZE(ins)));
-
-			/*
-			 * Turn off prefix and suffix compression until a full
-			 * key is written into the new page.
-			 */
-			r->key_pfx_compress = r->key_sfx_compress = false;
-			continue;
-		}
-
-		switch (upd->type) {
-		case WT_UPDATE_MODIFY:
-			/*
-			 * Impossible slot, there's no backing on-page
-			 * item.
-			 */
-			cbt->slot = UINT32_MAX;
-			WT_RET(__wt_value_return_upd(
-			    session, cbt, upd, F_ISSET(r, WT_REC_VISIBLE_ALL)));
-			WT_RET(__rec_cell_build_val(session, r,
-			    cbt->iface.value.data,
-			    cbt->iface.value.size, (uint64_t)0));
-			break;
-		case WT_UPDATE_STANDARD:
-			if (upd->size == 0)
-				val->len = 0;
-			else
-				WT_RET(__rec_cell_build_val(session,
-				    r, upd->data, upd->size,
-				    (uint64_t)0));
-			break;
-		case WT_UPDATE_TOMBSTONE:
-			continue;
-		WT_ILLEGAL_VALUE(session, upd->type);
-		}
-
-		/* Build key cell. */
-		WT_RET(__rec_cell_build_leaf_key(session, r,
-		    WT_INSERT_KEY(ins), WT_INSERT_KEY_SIZE(ins), &ovfl_key));
-
-		/* Boundary: split or write the page. */
-		if (__rec_need_split(r, key->len + val->len)) {
-			/*
-			 * Turn off prefix compression until a full key written
-			 * to the new page, and (unless already working with an
-			 * overflow key), rebuild the key without compression.
-			 */
-			if (r->key_pfx_compress_conf) {
-				r->key_pfx_compress = false;
-				if (!ovfl_key)
-					WT_RET(__rec_cell_build_leaf_key(
-					    session, r, NULL, 0, &ovfl_key));
-			}
-
-			WT_RET(__rec_split_crossing_bnd(
-			    session, r, key->len + val->len));
-		}
-
-		/* Copy the key/value pair onto the page. */
-		__rec_copy_incr(session, r, key);
-		if (val->len == 0)
-			r->any_empty_value = true;
-		else {
-			r->all_empty_value = false;
-			if (btree->dictionary)
-				WT_RET(__rec_dict_replace(session, r, 0, val));
-			__rec_copy_incr(session, r, val);
-		}
-
-		/* Update compression state. */
-		__rec_key_state_update(r, ovfl_key);
-	}
 
 	return (0);
 }
@@ -5649,232 +2511,12 @@ __rec_las_wrapup_err(WT_SESSION_IMPL *session, WT_RECONCILE *r)
 }
 
 /*
- * __rec_cell_build_int_key --
- *	Process a key and return a WT_CELL structure and byte string to be
- *	stored on a row-store internal page.
- */
-static int
-__rec_cell_build_int_key(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, const void *data, size_t size, bool *is_ovflp)
-{
-	WT_BTREE *btree;
-	WT_KV *key;
-
-	*is_ovflp = false;
-
-	btree = S2BT(session);
-
-	key = &r->k;
-
-	/* Copy the bytes into the "current" and key buffers. */
-	WT_RET(__wt_buf_set(session, r->cur, data, size));
-	WT_RET(__wt_buf_set(session, &key->buf, data, size));
-
-	/* Create an overflow object if the data won't fit. */
-	if (size > btree->maxintlkey) {
-		WT_STAT_DATA_INCR(session, rec_overflow_key_internal);
-
-		*is_ovflp = true;
-		return (__rec_cell_build_ovfl(
-		    session, r, key, WT_CELL_KEY_OVFL, (uint64_t)0));
-	}
-
-	key->cell_len = __wt_cell_pack_int_key(&key->cell, key->buf.size);
-	key->len = key->cell_len + key->buf.size;
-
-	return (0);
-}
-
-/*
- * __rec_cell_build_leaf_key --
- *	Process a key and return a WT_CELL structure and byte string to be
- *	stored on a row-store leaf page.
- */
-static int
-__rec_cell_build_leaf_key(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, const void *data, size_t size, bool *is_ovflp)
-{
-	WT_BTREE *btree;
-	WT_KV *key;
-	size_t pfx_max;
-	const uint8_t *a, *b;
-	uint8_t pfx;
-
-	*is_ovflp = false;
-
-	btree = S2BT(session);
-
-	key = &r->k;
-
-	pfx = 0;
-	if (data == NULL)
-		/*
-		 * When data is NULL, our caller has a prefix compressed key
-		 * they can't use (probably because they just crossed a split
-		 * point).  Use the full key saved when last called, instead.
-		 */
-		WT_RET(__wt_buf_set(
-		    session, &key->buf, r->cur->data, r->cur->size));
-	else {
-		/*
-		 * Save a copy of the key for later reference: we use the full
-		 * key for prefix-compression comparisons, and if we are, for
-		 * any reason, unable to use the compressed key we generate.
-		 */
-		WT_RET(__wt_buf_set(session, r->cur, data, size));
-
-		/*
-		 * Do prefix compression on the key.  We know by definition the
-		 * previous key sorts before the current key, which means the
-		 * keys must differ and we just need to compare up to the
-		 * shorter of the two keys.
-		 */
-		if (r->key_pfx_compress) {
-			/*
-			 * We can't compress out more than 256 bytes, limit the
-			 * comparison to that.
-			 */
-			pfx_max = UINT8_MAX;
-			if (size < pfx_max)
-				pfx_max = size;
-			if (r->last->size < pfx_max)
-				pfx_max = r->last->size;
-			for (a = data, b = r->last->data; pfx < pfx_max; ++pfx)
-				if (*a++ != *b++)
-					break;
-
-			/*
-			 * Prefix compression may cost us CPU and memory when
-			 * the page is re-loaded, don't do it unless there's
-			 * reasonable gain.
-			 */
-			if (pfx < btree->prefix_compression_min)
-				pfx = 0;
-			else
-				WT_STAT_DATA_INCRV(
-				    session, rec_prefix_compression, pfx);
-		}
-
-		/* Copy the non-prefix bytes into the key buffer. */
-		WT_RET(__wt_buf_set(
-		    session, &key->buf, (uint8_t *)data + pfx, size - pfx));
-	}
-
-	/* Optionally compress the key using the Huffman engine. */
-	if (btree->huffman_key != NULL)
-		WT_RET(__wt_huffman_encode(session, btree->huffman_key,
-		    key->buf.data, (uint32_t)key->buf.size, &key->buf));
-
-	/* Create an overflow object if the data won't fit. */
-	if (key->buf.size > btree->maxleafkey) {
-		/*
-		 * Overflow objects aren't prefix compressed -- rebuild any
-		 * object that was prefix compressed.
-		 */
-		if (pfx == 0) {
-			WT_STAT_DATA_INCR(session, rec_overflow_key_leaf);
-
-			*is_ovflp = true;
-			return (__rec_cell_build_ovfl(
-			    session, r, key, WT_CELL_KEY_OVFL, (uint64_t)0));
-		}
-		return (
-		    __rec_cell_build_leaf_key(session, r, NULL, 0, is_ovflp));
-	}
-
-	key->cell_len = __wt_cell_pack_leaf_key(&key->cell, pfx, key->buf.size);
-	key->len = key->cell_len + key->buf.size;
-
-	return (0);
-}
-
-/*
- * __rec_cell_build_addr --
- *	Process an address reference and return a cell structure to be stored
- *	on the page.
- */
-static void
-__rec_cell_build_addr(WT_SESSION_IMPL *session, WT_RECONCILE *r,
-    const void *addr, size_t size, u_int cell_type, uint64_t recno)
-{
-	WT_KV *val;
-
-	val = &r->v;
-
-	WT_ASSERT(session, size != 0 || cell_type == WT_CELL_ADDR_DEL);
-
-	/*
-	 * We don't check the address size because we can't store an address on
-	 * an overflow page: if the address won't fit, the overflow page's
-	 * address won't fit either.  This possibility must be handled by Btree
-	 * configuration, we have to disallow internal page sizes that are too
-	 * small with respect to the largest address cookie the underlying block
-	 * manager might return.
-	 */
-
-	/*
-	 * We don't copy the data into the buffer, it's not necessary; just
-	 * re-point the buffer's data/length fields.
-	 */
-	val->buf.data = addr;
-	val->buf.size = size;
-	val->cell_len =
-	    __wt_cell_pack_addr(&val->cell, cell_type, recno, val->buf.size);
-	val->len = val->cell_len + val->buf.size;
-}
-
-/*
- * __rec_cell_build_val --
- *	Process a data item and return a WT_CELL structure and byte string to
- *	be stored on the page.
- */
-static int
-__rec_cell_build_val(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, const void *data, size_t size, uint64_t rle)
-{
-	WT_BTREE *btree;
-	WT_KV *val;
-
-	btree = S2BT(session);
-
-	val = &r->v;
-
-	/*
-	 * We don't copy the data into the buffer, it's not necessary; just
-	 * re-point the buffer's data/length fields.
-	 */
-	val->buf.data = data;
-	val->buf.size = size;
-
-	/* Handle zero-length cells quickly. */
-	if (size != 0) {
-		/* Optionally compress the data using the Huffman engine. */
-		if (btree->huffman_value != NULL)
-			WT_RET(__wt_huffman_encode(
-			    session, btree->huffman_value,
-			    val->buf.data, (uint32_t)val->buf.size, &val->buf));
-
-		/* Create an overflow object if the data won't fit. */
-		if (val->buf.size > btree->maxleafvalue) {
-			WT_STAT_DATA_INCR(session, rec_overflow_value);
-
-			return (__rec_cell_build_ovfl(
-			    session, r, val, WT_CELL_VALUE_OVFL, rle));
-		}
-	}
-	val->cell_len = __wt_cell_pack_data(&val->cell, rle, val->buf.size);
-	val->len = val->cell_len + val->buf.size;
-
-	return (0);
-}
-
-/*
- * __rec_cell_build_ovfl --
+ * __wt_rec_cell_build_ovfl --
  *	Store overflow items in the file, returning the address cookie.
  */
-static int
-__rec_cell_build_ovfl(WT_SESSION_IMPL *session,
-    WT_RECONCILE *r, WT_KV *kv, uint8_t type, uint64_t rle)
+int
+__wt_rec_cell_build_ovfl(WT_SESSION_IMPL *session,
+    WT_RECONCILE *r, WT_REC_KV *kv, uint8_t type, uint64_t rle)
 {
 	WT_BM *bm;
 	WT_BTREE *btree;
@@ -5938,195 +2580,4 @@ __rec_cell_build_ovfl(WT_SESSION_IMPL *session,
 
 err:	__wt_scr_free(session, &tmp);
 	return (ret);
-}
-
-/*
- * __rec_dictionary_skip_search --
- *	Search a dictionary skiplist.
- */
-static WT_DICTIONARY *
-__rec_dictionary_skip_search(WT_DICTIONARY **head, uint64_t hash)
-{
-	WT_DICTIONARY **e;
-	int i;
-
-	/*
-	 * Start at the highest skip level, then go as far as possible at each
-	 * level before stepping down to the next.
-	 */
-	for (i = WT_SKIP_MAXDEPTH - 1, e = &head[i]; i >= 0;) {
-		if (*e == NULL) {		/* Empty levels */
-			--i;
-			--e;
-			continue;
-		}
-
-		/*
-		 * Return any exact matches: we don't care in what search level
-		 * we found a match.
-		 */
-		if ((*e)->hash == hash)		/* Exact match */
-			return (*e);
-		if ((*e)->hash > hash) {	/* Drop down a level */
-			--i;
-			--e;
-		} else				/* Keep going at this level */
-			e = &(*e)->next[i];
-	}
-	return (NULL);
-}
-
-/*
- * __rec_dictionary_skip_search_stack --
- *	Search a dictionary skiplist, returning an insert/remove stack.
- */
-static void
-__rec_dictionary_skip_search_stack(
-    WT_DICTIONARY **head, WT_DICTIONARY ***stack, uint64_t hash)
-{
-	WT_DICTIONARY **e;
-	int i;
-
-	/*
-	 * Start at the highest skip level, then go as far as possible at each
-	 * level before stepping down to the next.
-	 */
-	for (i = WT_SKIP_MAXDEPTH - 1, e = &head[i]; i >= 0;)
-		if (*e == NULL || (*e)->hash > hash)
-			stack[i--] = e--;	/* Drop down a level */
-		else
-			e = &(*e)->next[i];	/* Keep going at this level */
-}
-
-/*
- * __rec_dictionary_skip_insert --
- *	Insert an entry into the dictionary skip-list.
- */
-static void
-__rec_dictionary_skip_insert(
-    WT_DICTIONARY **head, WT_DICTIONARY *e, uint64_t hash)
-{
-	WT_DICTIONARY **stack[WT_SKIP_MAXDEPTH];
-	u_int i;
-
-	/* Insert the new entry into the skiplist. */
-	__rec_dictionary_skip_search_stack(head, stack, hash);
-	for (i = 0; i < e->depth; ++i) {
-		e->next[i] = *stack[i];
-		*stack[i] = e;
-	}
-}
-
-/*
- * __rec_dictionary_init --
- *	Allocate and initialize the dictionary.
- */
-static int
-__rec_dictionary_init(WT_SESSION_IMPL *session, WT_RECONCILE *r, u_int slots)
-{
-	u_int depth, i;
-
-	/* Free any previous dictionary. */
-	__rec_dictionary_free(session, r);
-
-	r->dictionary_slots = slots;
-	WT_RET(__wt_calloc(session,
-	    r->dictionary_slots, sizeof(WT_DICTIONARY *), &r->dictionary));
-	for (i = 0; i < r->dictionary_slots; ++i) {
-		depth = __wt_skip_choose_depth(session);
-		WT_RET(__wt_calloc(session, 1,
-		    sizeof(WT_DICTIONARY) + depth * sizeof(WT_DICTIONARY *),
-		    &r->dictionary[i]));
-		r->dictionary[i]->depth = depth;
-	}
-	return (0);
-}
-
-/*
- * __rec_dictionary_free --
- *	Free the dictionary.
- */
-static void
-__rec_dictionary_free(WT_SESSION_IMPL *session, WT_RECONCILE *r)
-{
-	u_int i;
-
-	if (r->dictionary == NULL)
-		return;
-
-	/*
-	 * We don't correct dictionary_slots when we fail during allocation,
-	 * but that's OK, the value is either NULL or a memory reference to
-	 * be free'd.
-	 */
-	for (i = 0; i < r->dictionary_slots; ++i)
-		__wt_free(session, r->dictionary[i]);
-	__wt_free(session, r->dictionary);
-}
-
-/*
- * __rec_dictionary_reset --
- *	Reset the dictionary when reconciliation restarts and when crossing a
- *	page boundary (a potential split).
- */
-static void
-__rec_dictionary_reset(WT_RECONCILE *r)
-{
-	if (r->dictionary_slots) {
-		r->dictionary_next = 0;
-		memset(r->dictionary_head, 0, sizeof(r->dictionary_head));
-	}
-}
-
-/*
- * __rec_dictionary_lookup --
- *	Check the dictionary for a matching value on this page.
- */
-static int
-__rec_dictionary_lookup(
-    WT_SESSION_IMPL *session, WT_RECONCILE *r, WT_KV *val, WT_DICTIONARY **dpp)
-{
-	WT_DICTIONARY *dp, *next;
-	uint64_t hash;
-	bool match;
-
-	*dpp = NULL;
-
-	/* Search the dictionary, and return any match we find. */
-	hash = __wt_hash_fnv64(val->buf.data, val->buf.size);
-	for (dp = __rec_dictionary_skip_search(r->dictionary_head, hash);
-	    dp != NULL && dp->hash == hash; dp = dp->next[0]) {
-		WT_RET(__wt_cell_pack_data_match(
-		    (WT_CELL *)((uint8_t *)r->cur_ptr->image.mem + dp->offset),
-		    &val->cell, val->buf.data, &match));
-		if (match) {
-			WT_STAT_DATA_INCR(session, rec_dictionary);
-			*dpp = dp;
-			return (0);
-		}
-	}
-
-	/*
-	 * We're not doing value replacement in the dictionary.  We stop adding
-	 * new entries if we run out of empty dictionary slots (but continue to
-	 * use the existing entries).  I can't think of any reason a leaf page
-	 * value is more likely to be seen because it was seen more recently
-	 * than some other value: if we find working sets where that's not the
-	 * case, it shouldn't be too difficult to maintain a pointer which is
-	 * the next dictionary slot to re-use.
-	 */
-	if (r->dictionary_next >= r->dictionary_slots)
-		return (0);
-
-	/*
-	 * Set the hash value, we'll add this entry into the dictionary when we
-	 * write it into the page's disk image buffer (because that's when we
-	 * know where on the page it will be written).
-	 */
-	next = r->dictionary[r->dictionary_next++];
-	next->offset = 0;		/* Not necessary, just cautious. */
-	next->hash = hash;
-	__rec_dictionary_skip_insert(r->dictionary_head, next, hash);
-	*dpp = next;
-	return (0);
 }
